chore(core): clean up swap v1
