+++ mobile-app/app/screens/AppNavigator/screens/Dex/CompositeSwap/CompositeSwapScreen.tsx
@@ -1,15 +1,14 @@
-import { useCallback, useEffect, useRef, useState } from "react";
-import { Platform, TouchableOpacity, View } from "react-native";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { Platform, ScrollView, TextInput, View } from "react-native";
 import { useSelector } from "react-redux";
-import { Control, Controller, useForm } from "react-hook-form";
+import { Controller, useForm } from "react-hook-form";
 import BigNumber from "bignumber.js";
 import {
   NavigationProp,
   useIsFocused,
   useNavigation,
 } from "@react-navigation/native";
-import { tailwind } from "@tailwind";
-import { BottomSheetModal } from "@gorhom/bottom-sheet";
+import { getColor, tailwind } from "@tailwind";
 import { translate } from "@translations";
 import { RootState } from "@store";
 import { hasTxQueued as hasBroadcastQueued } from "@store/ocean";
@@ -32,70 +31,71 @@ import { PoolPairData } from "@defichain/whale-api-client/dist/api/poolpairs";
 import { StackScreenProps } from "@react-navigation/stack";
 import {
   ThemedIcon,
-  ThemedScrollView,
-  ThemedText,
-  ThemedTouchableOpacity,
-  ThemedView,
+  ThemedScrollViewV2,
+  ThemedTextInputV2,
+  ThemedTextV2,
+  ThemedTouchableOpacityV2,
+  ThemedViewV2,
 } from "@components/themed";
-import { getNativeIcon } from "@components/icons/assets";
-import {
-  BottomSheetNavScreen,
-  BottomSheetWebWithNav,
-  BottomSheetWithNav,
-} from "@components/BottomSheetWithNav";
-import {
-  BottomSheetToken,
-  BottomSheetTokenList,
-  TokenType,
-} from "@components/BottomSheetTokenList";
-import { ConversionInfoText } from "@components/ConversionInfoText";
-import { InfoRow, InfoType } from "@components/InfoRow";
-import { InputHelperText } from "@components/InputHelperText";
-import { NumberRow } from "@components/NumberRow";
-import {
-  AmountButtonTypes,
-  SetAmountButton,
-} from "@components/SetAmountButton";
-import { WalletTextInput } from "@components/WalletTextInput";
-import { ReservedDFIInfoText } from "@components/ReservedDFIInfoText";
+import { BottomSheetToken } from "@components/BottomSheetTokenList";
 import { useWalletContext } from "@shared-contexts/WalletContext";
-import { SubmitButtonGroup } from "@components/SubmitButtonGroup";
 import { useTokenPrice } from "@screens/AppNavigator/screens/Portfolio/hooks/TokenPrice";
-import { useDeFiScanContext } from "@shared-contexts/DeFiScanContext";
-import { openURL } from "@api/linking";
-import NumberFormat from "react-number-format";
-import { TextRow } from "@components/TextRow";
-import { PriceRateProps, PricesSection } from "@components/PricesSection";
+
 import { fetchExecutionBlock } from "@store/futureSwap";
 import { useAppDispatch } from "@hooks/useAppDispatch";
 import { WalletAlert } from "@components/WalletAlert";
 import { useFeatureFlagContext } from "@contexts/FeatureFlagContext";
-import { AnnouncementBanner } from "../../Portfolio/components/Announcements";
+import { useThemeContext } from "@shared-contexts/ThemeProvider";
+import { useBottomSheet } from "@hooks/useBottomSheet";
 import {
-  DexStabilizationType,
-  useDexStabilization,
-} from "../hook/DexStabilization";
+  BottomSheetNavScreen,
+  BottomSheetWebWithNavV2,
+  BottomSheetWithNavV2,
+} from "@components/BottomSheetWithNavV2";
+import { PriceRateProps } from "@components/PricesSectionV2";
+import { SubmitButtonGroupV2 } from "@components/SubmitButtonGroupV2";
+import { TokenListType } from "@screens/AppNavigator/screens/Dex/CompositeSwap/SwapTokenSelectionScreen";
+import { useSwappableTokens } from "@screens/AppNavigator/screens/Dex/hook/SwappableTokens";
+import {
+  AmountButtonTypes,
+  TransactionCard,
+} from "@components/TransactionCard";
+import { useToast } from "react-native-toast-notifications";
+import { useDisplayUtxoWarning } from "@hooks/wallet/DisplayUtxoWarning";
+import { DEXIcon as SwapTokenButton } from "@screens/WalletNavigator/assets/BottomNavIcon";
+import {
+  Announcement,
+  AnnouncementBannerV2,
+} from "../../Portfolio/components/Announcements";
+import { useDexStabilization } from "../hook/DexStabilization";
 import { useFutureSwap, useFutureSwapDate } from "../hook/FutureSwap";
-import { ButtonGroup } from "../components/ButtonGroup";
-import { useSwappableTokens } from "../hook/SwappableTokens";
-import { useSlippageTolerance } from "../hook/SlippageTolerance";
 import { useTokenBestPath } from "../../Portfolio/hooks/TokenBestPath";
 import { DexParamList } from "../DexNavigator";
+import {
+  ButtonGroupTabKey,
+  SwapButtonGroup,
+} from "./components/SwapButtonGroup";
+import {
+  TokenDropdownButton,
+  TokenDropdownButtonStatus,
+} from "./components/TokenDropdownButton";
+import { ActiveUSDValueV2 } from "../../Loans/VaultDetail/components/ActiveUSDValueV2";
 import {
   SlippageError,
   SlippageTolerance,
 } from "./components/SlippageTolerance";
-
-export enum ButtonGroupTabKey {
-  InstantSwap = "INSTANT_SWAP",
-  FutureSwap = "FUTURE_SWAP",
-}
+import { BottomSheetSlippageInfo } from "./components/BottomSheetSlippageInfo";
+import { FutureSwapRowTo, InstantSwapRowTo } from "./components/SwapRowTo";
+import { SwapSummary } from "./components/SwapSummary";
+import { getPrecisedCurrencyValue } from "../../Auctions/helpers/precision-token-value";
+import { useSlippageTolerance } from "../hook/SlippageTolerance";
 
 export interface TokenState {
   id: string;
   reserve: string;
   displaySymbol: string;
   symbol: string;
+  name?: string;
 }
 
 export interface OwnedTokenState extends TokenState {
@@ -109,11 +109,15 @@ export function CompositeSwapScreen({ route }: Props): JSX.Element {
   const client = useWhaleApiClient();
   const whaleRpcClient = useWhaleRpcClient();
   const isFocused = useIsFocused();
+  const { isLight } = useThemeContext();
   const navigation = useNavigation<NavigationProp<DexParamList>>();
   const dispatch = useAppDispatch();
   const { address } = useWalletContext();
-  const { getArbitraryPoolPair, calculatePriceRates } = useTokenBestPath();
+  const { getArbitraryPoolPair, calculatePriceRates, getBestPath } =
+    useTokenBestPath();
+  const { getTokenPrice } = useTokenPrice();
   const { slippage, setSlippage } = useSlippageTolerance();
+  const { getDisplayUtxoWarningStatus } = useDisplayUtxoWarning();
 
   const blockCount = useSelector((state: RootState) => state.block.count ?? 0);
   const pairs = useSelector((state: RootState) => state.wallet.poolpairs);
@@ -145,26 +149,23 @@ export function CompositeSwapScreen({ route }: Props): JSX.Element {
   const [selectedTokenB, setSelectedTokenB] = useState<TokenState>();
   const [selectedPoolPairs, setSelectedPoolPairs] = useState<PoolPairData[]>();
   const [priceRates, setPriceRates] = useState<PriceRateProps[]>();
-  const [isModalDisplayed, setIsModalDisplayed] = useState(false);
-  const [isFromTokenSelectDisabled, setIsFromTokenSelectDisabled] =
-    useState(false);
-  const [isToTokenSelectDisabled, setIsToTokenSelectDisabled] = useState(false);
-  const buttonGroup = [
-    {
-      id: ButtonGroupTabKey.InstantSwap,
-      label: translate("screens/CompositeSwapScreen", "Instant Swap"),
-      handleOnPress: () => onButtonGroupChange(ButtonGroupTabKey.InstantSwap),
-    },
-    {
-      id: ButtonGroupTabKey.FutureSwap,
-      label: translate("screens/CompositeSwapScreen", "Future Swap"),
-      handleOnPress: () => onButtonGroupChange(ButtonGroupTabKey.FutureSwap),
-    },
-  ];
+  const [estimatedLessFeesAfterSlippage, setEstimatedLessFeesAfterSlippage] =
+    useState<BigNumber | undefined>(undefined);
   const [activeButtonGroup, setActiveButtonGroup] = useState<ButtonGroupTabKey>(
     ButtonGroupTabKey.InstantSwap
   );
   const [isFutureSwap, setIsFutureSwap] = useState(false);
+  const [bestPathEstimatedReturn, setBestPathEstimatedReturn] = useState<
+    { estimatedReturn: string; estimatedReturnLessDexFees: string } | undefined
+  >(undefined);
+  const [oraclePriceMessage, setOraclePriceMessage] = useState<string>(
+    translate(
+      "screens/CompositeSwapScreen",
+      "Future swap uses the oracle price of the selected token on the settlement block"
+    )
+  );
+  const [hasShownInputFocusBefore, setHasShownInputFocusBefore] =
+    useState<boolean>(false);
 
   const executionBlock = useSelector(
     (state: RootState) => state.futureSwaps.executionBlock
@@ -173,14 +174,32 @@ export function CompositeSwapScreen({ route }: Props): JSX.Element {
     executionBlock,
     blockCount
   );
-  const { fromTokens, toTokens } = useSwappableTokens(selectedTokenA?.id);
-  const { isFutureSwapOptionEnabled, oraclePriceText, isSourceLoanToken } =
-    useFutureSwap({
-      fromTokenDisplaySymbol: selectedTokenA?.displaySymbol,
-      toTokenDisplaySymbol: selectedTokenB?.displaySymbol,
-    });
-  const containerRef = useRef(null);
-  const bottomSheetRef = useRef<BottomSheetModal>(null);
+  const { fromTokens, toTokens } = useSwappableTokens(
+    selectedTokenA?.id,
+    selectedTokenA?.displaySymbol,
+    selectedTokenA?.symbol,
+    activeButtonGroup === ButtonGroupTabKey.FutureSwap
+  );
+  const {
+    isFutureSwapOptionEnabled,
+    oraclePriceText,
+    isSourceLoanToken,
+    isFromLoanToken,
+    isToLoanToken,
+  } = useFutureSwap({
+    fromTokenDisplaySymbol: selectedTokenA?.displaySymbol,
+    toTokenDisplaySymbol: selectedTokenB?.displaySymbol,
+  });
+  const amountInputRef = useRef<TextInput>();
+  const scrollViewRef = useRef<ScrollView>();
+
+  const {
+    bottomSheetRef,
+    containerRef,
+    dismissModal,
+    expandModal,
+    isModalDisplayed,
+  } = useBottomSheet();
 
   // dex stabilization
   const { isFeatureAvailable } = useFeatureFlagContext();
@@ -194,25 +213,26 @@ export function CompositeSwapScreen({ route }: Props): JSX.Element {
     },
   } = useDexStabilization(selectedTokenA, selectedTokenB);
 
-  const expandModal = useCallback(() => {
-    if (Platform.OS === "web") {
-      setIsModalDisplayed(true);
-    } else {
-      bottomSheetRef.current?.present();
-    }
-  }, []);
-
-  const dismissModal = useCallback(() => {
-    if (Platform.OS === "web") {
-      setIsModalDisplayed(false);
-    } else {
-      bottomSheetRef.current?.close();
-    }
-  }, []);
+  const toast = useToast();
+  const TOAST_DURATION = 2000;
+  const oraclePriceAnnouncement: Announcement = {
+    content: oraclePriceMessage,
+    url: "",
+    id: undefined,
+    type: "OTHER_ANNOUNCEMENT",
+    icon: "info",
+  };
 
   const onButtonGroupChange = (buttonGroupTabKey: ButtonGroupTabKey): void => {
     setActiveButtonGroup(buttonGroupTabKey);
+    scrollViewRef.current?.scrollTo({
+      y: 0,
+      animated: false,
+    });
   };
+
+  const [isEditingCustomSlippageInput, setIsEditingCustomSlippageInput] =
+    useState(false); // for custom slippage input
 
   // component UI state
   const { control, formState, setValue, trigger, watch } = useForm<{
@@ -220,6 +240,10 @@ export function CompositeSwapScreen({ route }: Props): JSX.Element {
     tokenB: string;
   }>({ mode: "onChange" });
   const { tokenA, tokenB } = watch();
+  const isReservedUtxoUsed = getDisplayUtxoWarningStatus(
+    new BigNumber(tokenA),
+    selectedTokenA?.displaySymbol ?? ""
+  );
 
   const { isConversionRequired, conversionAmount } = useConversion({
     inputToken: {
@@ -247,6 +271,14 @@ export function CompositeSwapScreen({ route }: Props): JSX.Element {
     { tokenId, reserve, token: { displaySymbol, symbol } }: BottomSheetToken,
     direction: "FROM" | "TO"
   ): void => {
+    if (
+      (selectedTokenA?.displaySymbol === displaySymbol &&
+        direction === "FROM") ||
+      (selectedTokenB?.displaySymbol === displaySymbol && direction === "TO")
+    ) {
+      return;
+    }
+
     const ownedToken = tokens?.find((token) => token.id === tokenId);
     const derivedToken = {
       id: ownedToken !== undefined ? ownedToken.id : tokenId, // retrieve unified token if selected
@@ -263,44 +295,45 @@ export function CompositeSwapScreen({ route }: Props): JSX.Element {
         ...derivedToken,
         amount: ownedToken === undefined ? "0" : ownedToken.amount,
       });
-      setActiveButtonGroup(ButtonGroupTabKey.InstantSwap);
-
-      if (selectedTokenB !== undefined) {
-        setSelectedTokenB(undefined);
-        setValue("tokenA", "");
-        setValue("tokenB", "");
-      }
+      setSelectedTokenB(undefined);
+      setValue("tokenA", "");
+      setValue("tokenB", "");
     } else {
       setSelectedTokenB(derivedToken);
     }
   };
 
-  const onBottomSheetSelect = ({
-    direction,
-  }: {
-    direction: "FROM" | "TO";
-  }): void => {
+  const BottomSheetHeader = {
+    headerStatusBarHeight: 2,
+    headerTitle: "",
+    headerBackTitleVisible: false,
+    headerStyle: tailwind("rounded-t-xl-v2 border-b-0", {
+      "bg-mono-light-v2-100": isLight,
+      "bg-mono-dark-v2-100": !isLight,
+    }),
+    headerRight: (): JSX.Element => {
+      return (
+        <ThemedTouchableOpacityV2
+          style={tailwind("mr-5", {
+            "mt-4 -mb-4": Platform.OS === "ios",
+            "mt-1.5": Platform.OS === "android",
+          })}
+          onPress={dismissModal}
+          testID="close_bottom_sheet_button"
+        >
+          <ThemedIcon iconType="Feather" name="x-circle" size={22} />
+        </ThemedTouchableOpacityV2>
+      );
+    },
+    headerLeft: () => <></>,
+  };
+
+  const onBottomSheetSlippageSelect = (): void => {
     setBottomSheetScreen([
       {
-        stackScreenName: "TokenList",
-        component: BottomSheetTokenList({
-          tokens: direction === "FROM" ? fromTokens ?? [] : toTokens ?? [],
-          tokenType: TokenType.BottomSheetToken,
-          headerLabel: translate(
-            "screens/CompositeSwapScreen",
-            direction === "FROM"
-              ? "Choose token for swap"
-              : "Choose token to swap"
-          ),
-          onCloseButtonPress: () => dismissModal(),
-          onTokenPress: (item): void => {
-            onTokenSelect(item, direction);
-            dismissModal();
-          },
-        }),
-        option: {
-          header: () => null,
-        },
+        stackScreenName: "SlippageInfo",
+        component: BottomSheetSlippageInfo(),
+        option: BottomSheetHeader,
       },
     ]);
     expandModal();
@@ -338,9 +371,6 @@ export function CompositeSwapScreen({ route }: Props): JSX.Element {
           isPreselected: true,
         },
       };
-
-    setIsFromTokenSelectDisabled(tokenSelectOption.from.isDisabled);
-    setIsToTokenSelectDisabled(tokenSelectOption.to.isDisabled);
 
     if (route.params.fromToken !== undefined) {
       onTokenSelect(
@@ -401,6 +431,15 @@ export function CompositeSwapScreen({ route }: Props): JSX.Element {
 
   useEffect(() => {
     void getSelectedPoolPairs();
+
+    if (hasShownInputFocusBefore || !isBothTokensSelected()) {
+      return;
+    }
+    /* timeout added to auto display keyboard on Android */
+    Platform.OS === "android"
+      ? setTimeout(() => amountInputRef?.current?.focus(), 0)
+      : amountInputRef?.current?.focus();
+    setHasShownInputFocusBefore(true);
   }, [selectedTokenA, selectedTokenB]);
 
   const getSelectedPoolPairs = async (): Promise<void> => {
@@ -424,11 +463,13 @@ export function CompositeSwapScreen({ route }: Props): JSX.Element {
       selectedPoolPairs !== undefined &&
       tokenA !== undefined
     ) {
-      const { aToBPrice, bToAPrice, estimated } = await calculatePriceRates(
-        selectedTokenA.id,
-        selectedTokenB.id,
-        new BigNumber(tokenA)
-      );
+      const { aToBPrice, bToAPrice, estimated, estimatedLessDexFees } =
+        await calculatePriceRates(
+          selectedTokenA.id,
+          selectedTokenB.id,
+          new BigNumber(tokenA)
+        );
+
       // Find the selected reserve in case it's null. From Token Detail Screen, reserve does not exist due to pair not existing
       const selectedReserve =
         selectedPoolPairs[0]?.tokenA?.id === selectedTokenA.id
@@ -441,28 +482,33 @@ export function CompositeSwapScreen({ route }: Props): JSX.Element {
       const slippage = new BigNumber(1).minus(
         new BigNumber(tokenA).div(tokenAReserve)
       );
+      setEstimatedLessFeesAfterSlippage(
+        new BigNumber(estimatedLessDexFees).multipliedBy(slippage)
+      );
 
-      const estimatedAmountAfterSlippage = estimated.times(slippage).toFixed(8);
       setPriceRates([
         {
-          label: translate("components/PricesSection", "1 {{token}}", {
-            token: selectedTokenB.displaySymbol,
-          }),
-          value: bToAPrice.toFixed(8),
-          aSymbol: selectedTokenB.displaySymbol,
-          bSymbol: selectedTokenA.displaySymbol,
-        },
-        {
-          label: translate("components/PricesSection", "1 {{token}}", {
+          label: translate("components/PricesSection", "1 {{token}} =", {
             token: selectedTokenA.displaySymbol,
           }),
           value: aToBPrice.toFixed(8),
+          symbolUSDValue: getAmountInUSDValue(selectedTokenA, new BigNumber(1)),
+          usdTextStyle: tailwind("text-sm"),
           aSymbol: selectedTokenA.displaySymbol,
           bSymbol: selectedTokenB.displaySymbol,
         },
+        {
+          label: translate("components/PricesSection", "1 {{token}} =", {
+            token: selectedTokenB.displaySymbol,
+          }),
+          value: bToAPrice.toFixed(8),
+          symbolUSDValue: getAmountInUSDValue(selectedTokenB, new BigNumber(1)),
+          usdTextStyle: tailwind("text-sm"),
+          aSymbol: selectedTokenB.displaySymbol,
+          bSymbol: selectedTokenA.displaySymbol,
+        },
       ]);
-
-      setValue("tokenB", estimatedAmountAfterSlippage);
+      setValue("tokenB", estimated.toFixed(8));
       // trigger validation for tokenB
       await trigger("tokenB");
     }
@@ -475,56 +521,100 @@ export function CompositeSwapScreen({ route }: Props): JSX.Element {
     );
   }, [activeButtonGroup, isFutureSwapOptionEnabled]);
 
-  const navigateToConfirmScreen = (): void => {
-    if (
-      selectedPoolPairs === undefined ||
-      selectedTokenA === undefined ||
-      selectedTokenB === undefined ||
-      priceRates === undefined ||
-      tokenA === undefined ||
-      tokenB === undefined
-    ) {
-      return;
+  useEffect(() => {
+    if (selectedTokenA === undefined || selectedTokenB === undefined) {
+      return undefined;
     }
 
-    const ownedTokenB = tokens.find((token) => token.id === selectedTokenB.id);
-    const slippageInDecimal = new BigNumber(slippage).div(100);
-    navigation.navigate("ConfirmCompositeSwapScreen", {
-      fee,
-      pairs: selectedPoolPairs,
-      priceRates,
-      slippage: slippageInDecimal,
-      futureSwap: isFutureSwap
-        ? {
-            executionBlock,
-            transactionDate,
-            isSourceLoanToken: isSourceLoanToken,
-            oraclePriceText,
+    const fetchBestPath = async () => {
+      const bestPath = await getBestPath(
+        selectedTokenA.id === "0_unified" ? "0" : selectedTokenA.id,
+        selectedTokenB.id === "0_unified" ? "0" : selectedTokenB.id
+      );
+      setBestPathEstimatedReturn({
+        estimatedReturn: bestPath.estimatedReturn,
+        estimatedReturnLessDexFees: bestPath.estimatedReturnLessDexFees,
+      });
+    };
+
+    fetchBestPath();
+  }, [selectedTokenA, selectedTokenB, selectedPoolPairs]);
+
+  const totalFees = useMemo(() => {
+    if (
+      tokenA === "" ||
+      new BigNumber(tokenA).isZero() ||
+      priceRates === undefined ||
+      selectedTokenB === undefined
+    ) {
+      return "-";
+    }
+
+    /*
+      dexFeesInTokenBUnit = Burn fees + commission fee of 1 tokenA
+    */
+    const dexFeesInTokenBUnit = new BigNumber(
+      bestPathEstimatedReturn?.estimatedReturn ?? 0
+    )
+      .minus(
+        new BigNumber(bestPathEstimatedReturn?.estimatedReturnLessDexFees ?? 0)
+      )
+      .multipliedBy(tokenA);
+
+    /* Transaction fee + DEX fees */
+    return getPrecisedCurrencyValue(
+      getTokenPrice("DFI", fee).plus(
+        getTokenPrice(selectedTokenB.symbol, dexFeesInTokenBUnit)
+      )
+    );
+  }, [priceRates, selectedTokenB, tokenA, bestPathEstimatedReturn, fee]);
+
+  useEffect(() => {
+    let message = translate(
+      "screens/CompositeSwapScreen",
+      "Future swap uses the oracle price of the selected token on the settlement block"
+    );
+    if (selectedTokenA !== undefined) {
+      if (selectedTokenA.displaySymbol === "DUSD") {
+        if (selectedTokenB === undefined) {
+          message = translate(
+            "screens/CompositeSwapScreen",
+            "You are buying dTokens at 5% more than the oracle price at settlement block."
+          );
+        } else {
+          message = translate(
+            "screens/CompositeSwapScreen",
+            "You are buying {{displaySymbol}} at 5% more than the oracle price at settlement block.",
+            { displaySymbol: selectedTokenB.displaySymbol }
+          );
+        }
+      } else {
+        message = translate(
+          "screens/CompositeSwapScreen",
+          "You are selling your {{displaySymbol}} at 5% less than the oracle price at settlement block.",
+          {
+            displaySymbol: selectedTokenA.displaySymbol,
           }
-        : undefined,
-      swap: {
-        tokenTo: selectedTokenB,
-        tokenFrom: selectedTokenA,
-        amountFrom: new BigNumber(tokenA),
-        amountTo: new BigNumber(tokenB),
+        );
+      }
+    }
+    setOraclePriceMessage(message);
+  }, [selectedTokenA, selectedTokenB]);
+
+  const navigateToTokenSelectionScreen = (listType: TokenListType): void => {
+    navigation.navigate("SwapTokenSelectionScreen", {
+      fromToken: {
+        symbol: selectedTokenA?.symbol,
+        displaySymbol: selectedTokenA?.displaySymbol,
       },
-      tokenA: selectedTokenA,
-      tokenB:
-        ownedTokenB !== undefined
-          ? {
-              ...selectedTokenB,
-              amount: ownedTokenB.amount,
-            }
-          : selectedTokenB,
-      ...(isConversionRequired && {
-        conversion: {
-          isConversionRequired,
-          DFIToken,
-          DFIUtxo,
-          conversionAmount,
-        },
-      }),
-      estimatedAmount: new BigNumber(tokenB),
+      listType: listType,
+      list: listType === TokenListType.From ? fromTokens ?? [] : toTokens ?? [],
+      onTokenPress: (item) => {
+        onTokenSelect(item, listType);
+        navigation.goBack();
+      },
+      isFutureSwap: activeButtonGroup === ButtonGroupTabKey.FutureSwap,
+      isSearchDTokensOnly: selectedTokenA?.displaySymbol === "DUSD",
     });
   };
 
@@ -567,6 +657,61 @@ export function CompositeSwapScreen({ route }: Props): JSX.Element {
     if (hasPendingJob || hasPendingBroadcastJob) {
       return;
     }
+
+    if (
+      selectedPoolPairs === undefined ||
+      selectedTokenA === undefined ||
+      selectedTokenB === undefined ||
+      priceRates === undefined ||
+      tokenA === undefined ||
+      tokenB === undefined
+    ) {
+      return;
+    }
+
+    const ownedTokenB = tokens.find((token) => token.id === selectedTokenB.id);
+    const slippageInDecimal = new BigNumber(slippage).div(100);
+    const params: DexParamList["ConfirmCompositeSwapScreen"] = {
+      fee,
+      pairs: selectedPoolPairs,
+      priceRates,
+      slippage: slippageInDecimal,
+      futureSwap: isFutureSwap
+        ? {
+            executionBlock,
+            transactionDate,
+            isSourceLoanToken: isSourceLoanToken,
+            oraclePriceText,
+          }
+        : undefined,
+      swap: {
+        tokenTo: selectedTokenB,
+        tokenFrom: selectedTokenA,
+        amountFrom: new BigNumber(tokenA),
+        amountTo: new BigNumber(tokenB),
+      },
+      tokenA: selectedTokenA,
+      tokenB:
+        ownedTokenB !== undefined
+          ? {
+              ...selectedTokenB,
+              amount: ownedTokenB.amount,
+            }
+          : selectedTokenB,
+      ...(isConversionRequired && {
+        conversion: {
+          isConversionRequired,
+          DFIToken,
+          DFIUtxo,
+          conversionAmount,
+        },
+      }),
+      estimatedAmount: new BigNumber(tokenB),
+      estimatedLessFeesAfterSlippage:
+        estimatedLessFeesAfterSlippage?.toFixed(8) ?? "",
+      totalFees,
+    };
+
     if (isConversionRequired) {
       queueConvertTransaction(
         {
@@ -575,12 +720,26 @@ export function CompositeSwapScreen({ route }: Props): JSX.Element {
         },
         dispatch,
         () => {
-          navigateToConfirmScreen();
+          navigation.navigate("ConfirmCompositeSwapScreen", params);
         },
-        logger
+        logger,
+        () => {
+          params.conversion = {
+            DFIUtxo,
+            DFIToken,
+            isConversionRequired: true,
+            conversionAmount,
+            isConverted: true,
+          };
+          navigation.navigate({
+            name: "ConfirmCompositeSwapScreen",
+            params,
+            merge: true,
+          });
+        }
       );
     } else {
-      navigateToConfirmScreen();
+      navigation.navigate("ConfirmCompositeSwapScreen", params);
     }
   };
 
@@ -602,254 +761,427 @@ export function CompositeSwapScreen({ route }: Props): JSX.Element {
     }
   };
 
+  const getAmountInUSDValue = useCallback(
+    (token: OwnedTokenState | TokenState | undefined, tokenAmount: any) => {
+      if (token === undefined || tokenAmount === "" || isNaN(tokenAmount)) {
+        return new BigNumber(0);
+      }
+
+      return getTokenPrice(token.symbol, new BigNumber(tokenAmount));
+    },
+    []
+  );
+
+  const isBothTokensSelected = (): boolean => {
+    return selectedTokenA !== undefined && selectedTokenB !== undefined;
+  };
+
+  async function onPercentagePress(
+    amount: string,
+    type: AmountButtonTypes
+  ): Promise<void> {
+    setValue("tokenA", amount);
+    await trigger("tokenA");
+    showToast(type);
+  }
+
+  function showToast(type: AmountButtonTypes): void {
+    if (selectedTokenA === undefined) {
+      return;
+    }
+
+    toast.hideAll();
+    const isMax = type === AmountButtonTypes.Max;
+    const toastMessage = isMax
+      ? "Max available {{unit}} entered"
+      : "{{percent}} of available {{unit}} entered";
+    const toastOption = {
+      unit: selectedTokenA.displaySymbol,
+      percent: type,
+    };
+    toast.show(
+      translate("screens/CompositeSwapScreen", toastMessage, toastOption),
+      {
+        type: "wallet_toast",
+        placement: "top",
+        duration: TOAST_DURATION,
+      }
+    );
+  }
+
   return (
     <View style={tailwind("h-full")} ref={containerRef}>
-      <ThemedScrollView>
-        {fromTokens !== undefined && fromTokens?.length > 0 && (
-          <ThemedText
-            dark={tailwind("text-gray-50")}
-            light={tailwind("text-gray-900")}
-            style={tailwind("text-xl font-semibold m-4 mb-0")}
-          >
-            {translate("screens/CompositeSwapScreen", "Swap tokens")}
-          </ThemedText>
-        )}
-
-        <View
-          style={tailwind([
-            "flex flex-row mt-3 mx-2",
-            { hidden: fromTokens?.length === 0 },
-          ])}
-        >
-          <TokenSelection
-            label={translate("screens/CompositeSwapScreen", "FROM")}
-            symbol={selectedTokenA?.displaySymbol}
-            onPress={() => onBottomSheetSelect({ direction: "FROM" })}
-            disabled={
-              isFromTokenSelectDisabled ||
-              fromTokens === undefined ||
-              fromTokens?.length === 0
-            }
-          />
-          <TokenSelection
-            label={translate("screens/CompositeSwapScreen", "TO")}
-            symbol={selectedTokenB?.displaySymbol}
-            onPress={() => onBottomSheetSelect({ direction: "TO" })}
-            disabled={
-              isToTokenSelectDisabled ||
-              toTokens === undefined ||
-              toTokens?.length === 0
-            }
-          />
-        </View>
-        {isDexStabilizationEnabled &&
-          dexStabilizationType !== "none" &&
-          dexStabilizationAnnouncement !== undefined && (
-            <View style={tailwind("flex mx-4 mt-4 rounded")}>
-              <AnnouncementBanner
-                announcement={dexStabilizationAnnouncement}
-                testID="swap_announcements_banner"
-              />
-            </View>
-          )}
-        <ThemedView
-          style={tailwind("m-4 pt-4 rounded-lg flex-1")}
-          light={tailwind("bg-white")}
-          dark={tailwind("bg-gray-800")}
-        >
-          {isFutureSwapOptionEnabled && (
-            <View style={tailwind("mb-4 mx-4")}>
-              <ButtonGroup
-                buttons={buttonGroup}
-                activeButtonGroupItem={activeButtonGroup}
-                labelStyle={tailwind("font-medium text-xs text-center py-0.5")}
-                testID="swap_button_group"
-                containerThemedProps={{
-                  light: tailwind("bg-gray-100"),
-                  dark: tailwind("bg-gray-900"),
-                }}
-              />
-            </View>
-          )}
-          {(selectedTokenA === undefined || selectedTokenB === undefined) &&
-            fromTokens?.length !== 0 && (
-              <ThemedText
-                dark={tailwind("text-gray-400")}
-                light={tailwind("text-gray-500")}
-                style={tailwind("text-center px-4 pb-4")}
-                testID="swap_instructions"
-              >
-                {" "}
-                {translate(
-                  "screens/CompositeSwapScreen",
-                  "Select tokens you want to swap to get started"
-                )}
-              </ThemedText>
-            )}
-
-          {selectedTokenA !== undefined && selectedTokenB !== undefined && (
-            <View style={tailwind("mx-4 flex-1")}>
-              <TokenRow
-                control={control}
-                title={translate(
-                  "screens/CompositeSwapScreen",
-                  "Enter amount to swap"
-                )}
-                maxAmount={getMaxAmount(selectedTokenA)}
-                onChangeFromAmount={async (amount) => {
-                  amount = isNaN(+amount) ? "0" : amount;
-                  setValue("tokenA", amount);
-                  await trigger("tokenA");
-                }}
-              />
-              <InputHelperText
-                testID="text_balance_amount"
-                label={`${translate(
-                  "screens/CompositeSwapScreen",
-                  "Available:"
-                )} `}
-                content={getMaxAmount(selectedTokenA)}
-                suffix={` ${selectedTokenA.displaySymbol}`}
-                labelStyleProps={tailwind("text-xs")}
-                styleProps={tailwind("text-xs")}
-              />
-              {selectedTokenA.id === "0_unified" && (
-                <View style={tailwind("mb-4")}>
-                  <ReservedDFIInfoText />
-                </View>
-              )}
-              <View
-                style={tailwind([
-                  "flex flex-row items-center",
-                  { "mb-4": isConversionRequired },
-                ])}
-              >
-                <TouchableOpacity
-                  onPress={onTokenSwitch}
-                  testID="switch_button"
-                >
-                  <ThemedIcon
-                    name="swap-vert"
-                    size={24}
-                    iconType="MaterialIcons"
-                    style={tailwind("w-8 mx-2")}
-                    dark={tailwind("text-darkprimary-500")}
-                    light={tailwind("text-primary-500")}
-                  />
-                </TouchableOpacity>
-                <View style={tailwind("flex-1")}>
-                  {isFutureSwap ? (
-                    <OraclePriceRow
-                      tokenDisplaySymbol={selectedTokenB.displaySymbol}
-                      oraclePriceText={oraclePriceText}
-                    />
-                  ) : (
-                    <TargetTokenRow control={control} token={selectedTokenB} />
-                  )}
-                </View>
-              </View>
-              {isConversionRequired && <ConversionInfoText />}
-            </View>
-          )}
-          {isFutureSwap && selectedTokenB !== undefined && (
-            <ThemedView
-              style={tailwind(
-                "flex flex-row py-2 px-4 mt-6 items-center rounded-t rounded-b-lg justify-between"
-              )}
-              light={tailwind("bg-blue-100")}
-              dark={tailwind("bg-darkblue-50")}
-            >
-              <ThemedText
-                style={tailwind("text-xs")}
-                light={tailwind("text-gray-500")}
-                dark={tailwind("text-gray-400")}
-                testID="future_swap_warning_text"
-              >
-                {oraclePriceText === "+5%"
-                  ? `${translate(
-                      "screens/CompositeSwapScreen",
-                      "By using future swap, you are buying {{tokenSymbol}} at 5% more than the oracle price at Settlement block",
-                      {
-                        tokenSymbol: selectedTokenB.displaySymbol,
-                      }
-                    )}`
-                  : `${translate(
-                      "screens/CompositeSwapScreen",
-                      "By using future swap, you are selling {{tokenSymbol}} at 5% lower than the oracle price at Settlement block",
-                      {
-                        tokenSymbol: selectedTokenA?.displaySymbol,
-                      }
-                    )}`}
-              </ThemedText>
-            </ThemedView>
-          )}
-
+      <ThemedViewV2 style={tailwind("h-full")}>
+        <SwapButtonGroup
+          activeButtonGroup={activeButtonGroup}
+          onPress={(type) => onButtonGroupChange(type)}
+          disableFutureSwap={
+            (isFromLoanToken !== undefined && !isFromLoanToken) ||
+            (isToLoanToken !== undefined && !isToLoanToken)
+          }
+        />
+        <ThemedScrollViewV2 ref={scrollViewRef}>
           {activeButtonGroup === ButtonGroupTabKey.InstantSwap &&
-            selectedTokenB !== undefined &&
-            selectedTokenA !== undefined && (
-              <SlippageTolerance
-                setSlippage={setSlippage}
-                slippageError={slippageError}
-                setSlippageError={setSlippageError}
-                slippage={slippage}
-              />
+            isDexStabilizationEnabled &&
+            dexStabilizationType !== "none" &&
+            dexStabilizationAnnouncement !== undefined && (
+              <View style={tailwind("flex mx-5 mt-8 rounded")}>
+                <AnnouncementBannerV2
+                  announcement={dexStabilizationAnnouncement}
+                  testID="swap_announcements_banner"
+                  containerStyle={{
+                    light: tailwind("bg-transparent"),
+                    dark: tailwind("bg-transparent"),
+                  }}
+                />
+              </View>
             )}
-        </ThemedView>
 
-        {selectedTokenB !== undefined &&
-          selectedTokenA !== undefined &&
-          priceRates !== undefined &&
-          tokenA !== undefined &&
-          tokenA !== "" &&
-          tokenB !== undefined && (
-            <>
-              {!isFutureSwap && (
-                <ThemedView
-                  style={tailwind("rounded-t-lg mx-4 py-2")}
-                  dark={tailwind("bg-gray-800 border-b border-gray-700")}
-                  light={tailwind("bg-white border-b border-gray-200")}
-                >
-                  <PricesSection
-                    testID="pricerate_value"
-                    priceRates={priceRates}
-                    isCompact
-                  />
-                </ThemedView>
-              )}
-              <TransactionDetailsSection
-                isFutureSwap={isFutureSwap}
-                conversionAmount={conversionAmount}
-                estimatedAmount={tokenB}
-                fee={fee}
-                isConversionRequired={isConversionRequired}
-                tokenA={selectedTokenA}
-                tokenB={selectedTokenB}
-                executionBlock={executionBlock}
-                timeRemaining={timeRemaining}
-                transactionDate={transactionDate}
-                oraclePriceText={oraclePriceText}
-                isDexStabilizationEnabled={isDexStabilizationEnabled}
-                dexStabilizationType={dexStabilizationType}
-                dexStabilizationFee={dexStabilizationFee}
+          {activeButtonGroup === ButtonGroupTabKey.FutureSwap && (
+            <View style={tailwind("flex mx-5 mt-8 rounded")}>
+              <AnnouncementBannerV2
+                announcement={oraclePriceAnnouncement}
+                testID="oracle_announcements_banner"
+                containerStyle={{
+                  light: tailwind("bg-transparent"),
+                  dark: tailwind("bg-transparent"),
+                }}
               />
+            </View>
+          )}
+
+          <ThemedTextV2
+            style={tailwind("mx-10 text-xs font-normal-v2 mt-8")}
+            light={tailwind("text-mono-light-v2-500")}
+            dark={tailwind("text-mono-dark-v2-500")}
+            testID="text_balance_amount"
+          >
+            {translate(
+              "screens/CompositeSwapScreen",
+              "I HAVE {{totalAmount}} {{token}}",
+              {
+                totalAmount:
+                  selectedTokenA != null ? getMaxAmount(selectedTokenA) : "",
+                token:
+                  selectedTokenA != null ? selectedTokenA.displaySymbol : "",
+              }
+            )}
+          </ThemedTextV2>
+
+          <View style={tailwind("mx-5")}>
+            <TransactionCard
+              maxValue={
+                new BigNumber(
+                  selectedTokenA != null ? getMaxAmount(selectedTokenA) : 0
+                )
+              }
+              onChange={onPercentagePress}
+              componentStyle={{
+                light: tailwind("bg-transparent"),
+                dark: tailwind("bg-transparent"),
+              }}
+              containerStyle={{
+                light: tailwind("bg-transparent"),
+                dark: tailwind("bg-transparent"),
+              }}
+              amountButtonsStyle={{
+                light: tailwind("bg-mono-light-v2-00"),
+                dark: tailwind("bg-mono-dark-v2-00"),
+                style: tailwind("mt-6 rounded-xl-v2"),
+              }}
+              disabled={
+                selectedTokenA === undefined ||
+                new BigNumber(getMaxAmount(selectedTokenA)).isZero()
+              }
+            >
+              <View
+                style={tailwind(
+                  "flex flex-row justify-between items-center pl-5 mt-4"
+                )}
+              >
+                <View style={tailwind("w-6/12 mr-2")}>
+                  <Controller
+                    control={control}
+                    defaultValue=""
+                    name="tokenA"
+                    render={({ field: { onChange, value } }) => (
+                      <ThemedTextInputV2
+                        style={tailwind("text-xl font-semibold-v2 w-full")}
+                        light={tailwind("text-mono-light-v2-900", {
+                          "opacity-30": selectedTokenA === undefined,
+                        })}
+                        dark={tailwind("text-mono-dark-v2-900", {
+                          "opacity-30": selectedTokenA === undefined,
+                        })}
+                        keyboardType="numeric"
+                        value={value}
+                        onBlur={async () => {
+                          await onChange(value?.trim());
+                        }}
+                        onChangeText={async (amount) => {
+                          amount = isNaN(+amount) ? "0" : amount;
+                          setValue("tokenA", amount);
+                          await trigger("tokenA");
+                        }}
+                        placeholder="0.00"
+                        placeholderTextColor={getColor(
+                          isLight ? "mono-light-v2-900" : "mono-dark-v2-900"
+                        )}
+                        ref={amountInputRef}
+                        testID="text_input_tokenA"
+                        editable={selectedTokenA !== undefined}
+                      />
+                    )}
+                    rules={{
+                      required: true,
+                      pattern: /^\d*\.?\d*$/,
+                      max: selectedTokenA ? getMaxAmount(selectedTokenA) : "0",
+                      validate: {
+                        greaterThanZero: (value: string) =>
+                          new BigNumber(
+                            value !== undefined && value !== "" ? value : 0
+                          ).isGreaterThan(0),
+                      },
+                    }}
+                  />
+                  <ActiveUSDValueV2
+                    price={getAmountInUSDValue(
+                      selectedTokenA ?? undefined,
+                      tokenA
+                    )}
+                    testId="tokenA_value_in_usd"
+                    containerStyle={tailwind("w-full break-words")}
+                  />
+                </View>
+
+                <TokenDropdownButton
+                  symbol={selectedTokenA?.displaySymbol}
+                  testID={TokenListType.From}
+                  onPress={() =>
+                    navigateToTokenSelectionScreen(TokenListType.From)
+                  }
+                  status={
+                    route.params.tokenSelectOption?.from?.isDisabled
+                      ? TokenDropdownButtonStatus.Locked
+                      : TokenDropdownButtonStatus.Enabled
+                  }
+                />
+              </View>
+            </TransactionCard>
+            <View style={tailwind("ml-5")}>
+              {tokenA !== "" && selectedTokenA === undefined && (
+                <ThemedTextV2
+                  light={tailwind("text-red-v2")}
+                  dark={tailwind("text-red-v2")}
+                  style={tailwind("text-xs pt-2 font-normal-v2")}
+                >
+                  {translate(
+                    "screens/CompositeSwapScreen",
+                    "Select a token first"
+                  )}
+                </ThemedTextV2>
+              )}
+              {formState.errors.tokenA?.type === "max" &&
+                selectedTokenA !== undefined &&
+                tokenA !== "" && (
+                  <ThemedTextV2
+                    light={tailwind("text-red-v2")}
+                    dark={tailwind("text-red-v2")}
+                    style={tailwind("text-xs pt-2 font-normal-v2")}
+                    testID="text_insufficient_balance"
+                  >
+                    {translate(
+                      "screens/CompositeSwapScreen",
+                      "Insufficient balance"
+                    )}
+                  </ThemedTextV2>
+                )}
+
+              {formState.errors.tokenA?.type === undefined &&
+                isReservedUtxoUsed && (
+                  <ThemedTextV2
+                    light={tailwind("text-orange-v2")}
+                    dark={tailwind("text-orange-v2")}
+                    style={tailwind("text-xs pt-2 font-normal-v2")}
+                    testID="utxo_reserved_fees_text"
+                  >
+                    {translate(
+                      "screens/CompositeSwapScreen",
+                      "A small amount of UTXO is reserved for fees"
+                    )}
+                  </ThemedTextV2>
+                )}
+            </View>
+
+            <View style={tailwind("my-8 flex-row")}>
+              <ThemedViewV2
+                dark={tailwind("border-mono-dark-v2-300")}
+                light={tailwind("border-mono-light-v2-300")}
+                style={tailwind("border-b-0.5 flex-1 h-1/2")}
+              />
+              <ThemedTouchableOpacityV2
+                onPress={onTokenSwitch}
+                style={tailwind("p-2.5 rounded-full z-50", {
+                  "opacity-30": !isBothTokensSelected(),
+                })}
+                dark={tailwind("bg-mono-dark-v2-900")}
+                light={tailwind("bg-mono-light-v2-900")}
+                testID="switch_button"
+                disabled={!isBothTokensSelected()}
+              >
+                <View
+                  style={{
+                    transform: [{ rotate: "90deg" }, { scaleX: -1 }],
+                  }}
+                >
+                  <SwapTokenButton
+                    size={24}
+                    color={
+                      isLight
+                        ? getColor("mono-light-v2-00")
+                        : getColor("mono-dark-v2-00")
+                    }
+                  />
+                </View>
+              </ThemedTouchableOpacityV2>
+              <ThemedViewV2
+                dark={tailwind("border-mono-dark-v2-300")}
+                light={tailwind("border-mono-light-v2-300")}
+                style={tailwind("border-b-0.5 flex-1 h-1/2")}
+              />
+            </View>
+
+            <ThemedViewV2
+              style={tailwind("border-0 pb-8")}
+              dark={tailwind("bg-transparent")}
+              light={tailwind("bg-transparent")}
+            >
+              <ThemedTextV2
+                style={tailwind("px-5 text-xs font-normal-v2")}
+                light={tailwind("text-mono-light-v2-500")}
+                dark={tailwind("text-mono-dark-v2-500")}
+                testID="tokenB_displaySymbol"
+              >
+                {translate("screens/CompositeSwapScreen", "I WANT {{token}}", {
+                  token: selectedTokenB?.displaySymbol ?? "",
+                })}
+              </ThemedTextV2>
+
+              <ThemedViewV2
+                style={tailwind(
+                  "flex flex-row justify-between items-center pl-5 mt-4",
+                  {
+                    "border-b-0.5 pb-8":
+                      isBothTokensSelected() &&
+                      priceRates !== undefined &&
+                      activeButtonGroup === ButtonGroupTabKey.InstantSwap,
+                  }
+                )}
+                light={tailwind("border-mono-light-v2-300")}
+                dark={tailwind("border-mono-dark-v2-300")}
+              >
+                {activeButtonGroup === ButtonGroupTabKey.FutureSwap && (
+                  <FutureSwapRowTo oraclePriceText={oraclePriceText} />
+                )}
+                {activeButtonGroup === ButtonGroupTabKey.InstantSwap && (
+                  <InstantSwapRowTo
+                    tokenAmount={new BigNumber(tokenB).toFixed(8)}
+                    tokenUsdAmount={getAmountInUSDValue(
+                      selectedTokenB ?? undefined,
+                      tokenB
+                    )}
+                  />
+                )}
+                <TokenDropdownButton
+                  symbol={selectedTokenB?.displaySymbol}
+                  testID={TokenListType.To}
+                  onPress={() =>
+                    navigateToTokenSelectionScreen(TokenListType.To)
+                  }
+                  status={
+                    route.params.tokenSelectOption?.to?.isDisabled
+                      ? TokenDropdownButtonStatus.Locked
+                      : selectedTokenA === undefined || toTokens.length === 0
+                      ? TokenDropdownButtonStatus.Disabled
+                      : TokenDropdownButtonStatus.Enabled
+                  }
+                />
+              </ThemedViewV2>
+            </ThemedViewV2>
+          </View>
+
+          {isBothTokensSelected() &&
+            activeButtonGroup === ButtonGroupTabKey.InstantSwap && (
+              <View style={tailwind("px-4 pb-6")}>
+                <SlippageTolerance
+                  setSlippage={setSlippage}
+                  setSlippageError={setSlippageError}
+                  onPress={onBottomSheetSlippageSelect}
+                  slippageError={slippageError}
+                  slippage={slippage}
+                  isEditing={isEditingCustomSlippageInput}
+                  setIsEditing={setIsEditingCustomSlippageInput}
+                />
+              </View>
+            )}
+
+          {isBothTokensSelected() && priceRates !== undefined && (
+            <>
+              <ThemedViewV2
+                light={tailwind("border-mono-light-v2-300")}
+                dark={tailwind("border-mono-dark-v2-300")}
+                style={tailwind("pt-5 px-5 mx-5 border-0.5 rounded-lg-v2")}
+              >
+                <SwapSummary
+                  instantSwapPriceRate={priceRates}
+                  activeTab={activeButtonGroup}
+                  executionBlock={executionBlock}
+                  transactionDate={transactionDate}
+                  transactionFee={fee}
+                  totalFees={totalFees}
+                  dexStabilizationFee={dexStabilizationFee}
+                  dexStabilizationType={dexStabilizationType}
+                  timeRemaining={timeRemaining}
+                />
+              </ThemedViewV2>
             </>
           )}
-        {selectedTokenA !== undefined && selectedTokenB !== undefined && (
-          <View style={tailwind("mb-2")}>
-            <SubmitButtonGroup
+
+          {formState.isValid && isBothTokensSelected() && (
+            <ThemedTextV2
+              testID="transaction_details_hint_text"
+              light={tailwind("text-mono-light-v2-500")}
+              dark={tailwind("text-mono-dark-v2-500")}
+              style={tailwind("pt-12 px-10 text-xs text-center font-normal-v2")}
+            >
+              {isConversionRequired
+                ? translate(
+                    "screens/CompositeSwapScreen",
+                    "By continuing, the required amount of DFI will be converted"
+                  )
+                : translate(
+                    "screens/CompositeSwapScreen",
+                    "Review full details in the next screen"
+                  )}
+            </ThemedTextV2>
+          )}
+
+          <View
+            style={tailwind("mb-12 mx-12 mt-16", {
+              "mt-5": formState.isValid && isBothTokensSelected(),
+            })}
+          >
+            <SubmitButtonGroupV2
               isDisabled={
                 !formState.isValid ||
                 hasPendingJob ||
                 hasPendingBroadcastJob ||
                 (slippageError?.type === "error" &&
                   slippageError !== undefined) ||
-                (isFutureSwap && isEnded)
+                (isFutureSwap && isEnded) ||
+                !isBothTokensSelected()
               }
-              label={translate("screens/CompositeSwapScreen", "CONTINUE")}
-              processingLabel={translate(
-                "screens/CompositeSwapScreen",
-                "CONTINUE"
-              )}
+              label={translate("components/Button", "Continue")}
               onSubmit={
                 (dexStabilizationType === "none" &&
                   isDexStabilizationEnabled) ||
@@ -858,527 +1190,42 @@ export function CompositeSwapScreen({ route }: Props): JSX.Element {
                   : onWarningBeforeSubmit
               }
               title="submit"
-              isProcessing={hasPendingJob || hasPendingBroadcastJob}
               displayCancelBtn={false}
+              buttonStyle="mt-0 mx-0"
             />
           </View>
-        )}
 
-        {formState.isValid &&
-          selectedTokenA !== undefined &&
-          selectedTokenB !== undefined && (
-            <ThemedText
-              testID="transaction_details_hint_text"
-              light={tailwind("text-gray-600")}
-              dark={tailwind("text-gray-300")}
-              style={tailwind("pb-8 px-4 text-sm text-center")}
-            >
-              {isConversionRequired
-                ? translate(
-                    "screens/CompositeSwapScreen",
-                    "Authorize transaction in the next screen to convert"
-                  )
-                : translate(
-                    "screens/CompositeSwapScreen",
-                    "Review and confirm transaction in the next screen"
-                  )}
-            </ThemedText>
+          {Platform.OS === "web" && (
+            <BottomSheetWebWithNavV2
+              modalRef={containerRef}
+              screenList={bottomSheetScreen}
+              isModalDisplayed={isModalDisplayed}
+              // eslint-disable-next-line react-native/no-inline-styles
+              modalStyle={{
+                position: "absolute",
+                bottom: "0",
+                height: "404px",
+                width: "375px",
+                zIndex: 50,
+                borderTopLeftRadius: 15,
+                borderTopRightRadius: 15,
+                overflow: "hidden",
+              }}
+            />
           )}
 
-        {Platform.OS === "web" && (
-          <BottomSheetWebWithNav
-            modalRef={containerRef}
-            screenList={bottomSheetScreen}
-            isModalDisplayed={isModalDisplayed}
-          />
-        )}
-
-        {Platform.OS !== "web" && (
-          <BottomSheetWithNav
-            modalRef={bottomSheetRef}
-            screenList={bottomSheetScreen}
-            snapPoints={{
-              ios: ["40%"],
-              android: ["45%"],
-            }}
-          />
-        )}
-      </ThemedScrollView>
-    </View>
-  );
-}
-
-function TokenSelection(props: {
-  symbol?: string;
-  label: string;
-  onPress: () => void;
-  disabled: boolean;
-}): JSX.Element {
-  const Icon = getNativeIcon(props.symbol ?? "");
-  return (
-    <View style={[tailwind("flex-grow mx-2"), { flexBasis: 0 }]}>
-      <ThemedText
-        dark={tailwind("text-gray-400")}
-        light={tailwind("text-gray-500")}
-        style={tailwind("text-xs pb-1")}
-      >
-        {props.label}
-      </ThemedText>
-      <ThemedTouchableOpacity
-        onPress={props.onPress}
-        testID={`token_select_button_${props.label}`}
-        dark={tailwind({
-          "bg-gray-600 text-gray-500 border-0": props.disabled,
-          "bg-gray-800 border-gray-400": !props.disabled,
-        })}
-        light={tailwind({
-          "bg-gray-200 border-0": props.disabled,
-          "bg-white border-gray-300": !props.disabled,
-        })}
-        style={tailwind("flex flex-row items-center border rounded p-2")}
-        disabled={props.disabled}
-      >
-        {props.symbol === undefined && (
-          <ThemedText
-            dark={tailwind("text-gray-400")}
-            light={tailwind("text-gray-500")}
-            style={tailwind("text-sm leading-6")}
-          >
-            {translate("screens/CompositeSwapScreen", "Select token")}
-          </ThemedText>
-        )}
-
-        {props.symbol !== undefined && (
-          <>
-            <Icon testID="tokenA_icon" height={17} width={17} />
-            <ThemedText
-              style={tailwind("ml-2")}
-              dark={tailwind({
-                "text-gray-200": !props.disabled,
-                "text-gray-400": props.disabled,
-              })}
-              light={tailwind({
-                "text-gray-900": !props.disabled,
-                "text-gray-500": props.disabled,
-              })}
-            >
-              {props.symbol}
-            </ThemedText>
-          </>
-        )}
-
-        <ThemedIcon
-          iconType="MaterialIcons"
-          name="unfold-more"
-          size={20}
-          dark={tailwind({
-            "text-darkprimary-500": !props.disabled,
-            "text-gray-400": props.disabled,
-          })}
-          light={tailwind({
-            "text-primary-500": !props.disabled,
-            "text-gray-500": props.disabled,
-          })}
-          style={[tailwind("text-center mt-0.5"), { marginLeft: "auto" }]}
-        />
-      </ThemedTouchableOpacity>
-    </View>
-  );
-}
-
-function TransactionDetailsSection({
-  isFutureSwap,
-  conversionAmount,
-  estimatedAmount,
-  fee,
-  isConversionRequired,
-  tokenA,
-  tokenB,
-  executionBlock,
-  timeRemaining,
-  transactionDate,
-  oraclePriceText,
-  isDexStabilizationEnabled,
-  dexStabilizationType,
-  dexStabilizationFee,
-}: {
-  isFutureSwap: boolean;
-  conversionAmount: BigNumber;
-  estimatedAmount: string;
-  fee: BigNumber;
-  isConversionRequired: boolean;
-  tokenA: OwnedTokenState;
-  tokenB: TokenState;
-  executionBlock: number;
-  timeRemaining: string;
-  transactionDate: string;
-  oraclePriceText: string;
-  isDexStabilizationEnabled: boolean;
-  dexStabilizationType: DexStabilizationType;
-  dexStabilizationFee?: string;
-}): JSX.Element {
-  const { getBlocksCountdownUrl } = useDeFiScanContext();
-  const { getTokenPrice } = useTokenPrice();
-  const rowStyle = {
-    lhsThemedProps: {
-      light: tailwind("text-gray-500"),
-      dark: tailwind("text-gray-400"),
-    },
-    rhsThemedProps: {
-      light: tailwind("text-gray-900"),
-      dark: tailwind("text-gray-50"),
-    },
-  };
-
-  return (
-    <View
-      style={tailwind("mx-4 overflow-hidden", {
-        "rounded-b-lg": !isFutureSwap,
-        "rounded-lg": isFutureSwap,
-      })}
-    >
-      {isConversionRequired && (
-        <NumberRow
-          lhs={translate("screens/CompositeSwapScreen", "UTXO to be converted")}
-          rhs={{
-            testID: "amount_to_convert",
-            value: conversionAmount.toFixed(8),
-            suffixType: "text",
-            suffix: tokenA.displaySymbol,
-          }}
-          lhsThemedProps={rowStyle.lhsThemedProps}
-          rhsThemedProps={rowStyle.rhsThemedProps}
-        />
-      )}
-
-      {!isFutureSwap ? (
-        <NumberRow
-          lhs={translate("screens/CompositeSwapScreen", "Estimated to receive")}
-          rhs={{
-            value: estimatedAmount,
-            suffixType: "text",
-            suffix: tokenB.displaySymbol,
-            testID: "estimated_to_receive",
-          }}
-          textStyle={tailwind("text-sm font-normal")}
-          rhsUsdAmount={getTokenPrice(
-            tokenB.symbol,
-            new BigNumber(estimatedAmount),
-            false
+          {Platform.OS !== "web" && (
+            <BottomSheetWithNavV2
+              modalRef={bottomSheetRef}
+              screenList={bottomSheetScreen}
+              snapPoints={{
+                ios: ["40%"],
+                android: ["45%"],
+              }}
+            />
           )}
-          lhsThemedProps={rowStyle.lhsThemedProps}
-          rhsThemedProps={rowStyle.rhsThemedProps}
-        />
-      ) : (
-        <>
-          <TimeRemainingTextRow
-            timeRemaining={timeRemaining}
-            transactionDate={transactionDate}
-          />
-          <InfoRow
-            type={InfoType.ExecutionBlock}
-            value={executionBlock}
-            testID="execution_block"
-            suffix={
-              <TouchableOpacity
-                onPress={async () =>
-                  await openURL(getBlocksCountdownUrl(executionBlock))
-                }
-              >
-                <ThemedIcon
-                  name="open-in-new"
-                  size={16}
-                  iconType="MaterialIcons"
-                  style={tailwind("ml-1")}
-                  light={tailwind("text-primary-500")}
-                  dark={tailwind("text-darkprimary-500")}
-                />
-              </TouchableOpacity>
-            }
-            lhsThemedProps={rowStyle.lhsThemedProps}
-            rhsThemedProps={rowStyle.rhsThemedProps}
-          />
-          <TextRow
-            lhs={{
-              value: translate(
-                "screens/ConfirmCompositeSwapScreen",
-                "Estimated to receive"
-              ),
-              themedProps: rowStyle.lhsThemedProps,
-              testID: "estimated_to_receive",
-            }}
-            rhs={{
-              value: translate(
-                "screens/CompositeSwapScreen",
-                `Oracle price ${oraclePriceText}`
-              ),
-              themedProps: rowStyle.rhsThemedProps,
-              testID: "estimated_to_receive",
-            }}
-            textStyle={tailwind("text-sm font-normal")}
-            containerStyle={{
-              dark: tailwind("bg-gray-800 border-b border-gray-700"),
-              light: tailwind("bg-white border-b border-gray-200"),
-              style: tailwind("p-4 flex-row items-start w-full"),
-            }}
-          />
-        </>
-      )}
-      <InfoRow
-        type={InfoType.EstimatedFee}
-        value={fee.toFixed(8)}
-        testID="text_fee"
-        suffix="DFI"
-        lhsThemedProps={rowStyle.lhsThemedProps}
-        rhsThemedProps={rowStyle.rhsThemedProps}
-      />
-      {isDexStabilizationEnabled &&
-        dexStabilizationType !== "none" &&
-        dexStabilizationFee !== undefined && (
-          <NumberRow
-            lhs={translate(
-              "screens/CompositeSwapScreen",
-              "DEX stabilization fee"
-            )}
-            rhs={{
-              value: dexStabilizationFee,
-              suffix: "%",
-              testID: "dex_stab_fee",
-              suffixType: "text",
-            }}
-            textStyle={tailwind("text-sm font-normal")}
-            lhsThemedProps={rowStyle.lhsThemedProps}
-            rhsThemedProps={rowStyle.rhsThemedProps}
-          />
-        )}
+        </ThemedScrollViewV2>
+      </ThemedViewV2>
     </View>
-  );
-}
-
-function TimeRemainingTextRow({
-  timeRemaining,
-  transactionDate,
-}: {
-  timeRemaining: string;
-  transactionDate: string;
-}): JSX.Element {
-  return (
-    <ThemedView
-      dark={tailwind("bg-gray-800 border-b border-gray-700")}
-      light={tailwind("bg-white border-b border-gray-200")}
-      style={tailwind("p-4 flex-row items-start w-full")}
-    >
-      <View style={tailwind("w-6/12")}>
-        <View style={tailwind("flex-row items-end justify-start")}>
-          <ThemedText
-            style={tailwind("text-sm")}
-            light={tailwind("text-gray-500")}
-            dark={tailwind("text-gray-400")}
-            testID="time_remaining_label"
-          >
-            {translate("screens/CompositeSwapScreen", "Est. time remaining")}
-          </ThemedText>
-        </View>
-      </View>
-      <View style={tailwind("flex flex-col justify-end flex-1")}>
-        <ThemedText
-          style={tailwind("text-sm text-right")}
-          light={tailwind("text-gray-900")}
-          dark={tailwind("text-gray-50")}
-          testID="time_remaining"
-        >
-          {`≈${timeRemaining}`}
-        </ThemedText>
-        <ThemedText
-          style={tailwind("text-xs text-right")}
-          light={tailwind("text-gray-500")}
-          dark={tailwind("text-gray-400")}
-        >
-          {`(${transactionDate})`}
-        </ThemedText>
-      </View>
-    </ThemedView>
-  );
-}
-
-interface TokenForm {
-  control: Control<{ tokenA: string; tokenB: string }>;
-  maxAmount?: string;
-  onChangeFromAmount: (amount: string) => void;
-  title: string;
-}
-
-function TokenRow(form: TokenForm): JSX.Element {
-  const { control, onChangeFromAmount, title, maxAmount } = form;
-  const rules: {
-    required: boolean;
-    pattern: RegExp;
-    validate: any;
-    max?: string;
-  } = {
-    required: true,
-    max: maxAmount,
-    pattern: /^\d*\.?\d*$/,
-    validate: {
-      greaterThanZero: (value: string) =>
-        new BigNumber(
-          value !== undefined && value !== "" ? value : 0
-        ).isGreaterThan(0),
-    },
-  };
-  const defaultValue = "";
-
-  return (
-    <Controller
-      control={control}
-      defaultValue={defaultValue}
-      name="tokenA"
-      render={({ field: { value } }) => (
-        <ThemedView
-          dark={tailwind("bg-transparent")}
-          light={tailwind("bg-transparent")}
-          style={tailwind("flex-row")}
-        >
-          <WalletTextInput
-            autoCapitalize="none"
-            onChange={(e) => {
-              onChangeFromAmount(e.nativeEvent.text);
-            }}
-            placeholder={translate(
-              "screens/CompositeSwapScreen",
-              "Enter an amount"
-            )}
-            style={tailwind("flex-grow w-2/5")}
-            testID="text_input_tokenA"
-            value={value}
-            displayClearButton={value !== defaultValue}
-            onClearButtonPress={() => onChangeFromAmount(defaultValue)}
-            title={title}
-            inputType="numeric"
-          >
-            <>
-              <SetAmountButton
-                amount={new BigNumber(maxAmount ?? "0")}
-                onPress={onChangeFromAmount}
-                type={AmountButtonTypes.half}
-              />
-
-              <SetAmountButton
-                amount={new BigNumber(maxAmount ?? "0")}
-                onPress={onChangeFromAmount}
-                type={AmountButtonTypes.max}
-              />
-            </>
-          </WalletTextInput>
-        </ThemedView>
-      )}
-      rules={rules}
-    />
-  );
-}
-
-interface OraclePriceRowProps {
-  oraclePriceText: string;
-  tokenDisplaySymbol: string;
-}
-
-function OraclePriceRow({
-  oraclePriceText,
-  tokenDisplaySymbol,
-}: OraclePriceRowProps): JSX.Element {
-  const Icon = getNativeIcon(tokenDisplaySymbol);
-
-  return (
-    <ThemedView
-      light={tailwind("bg-gray-50")}
-      style={tailwind(
-        "flex-row flex-grow justify-between items-center p-2 rounded"
-      )}
-    >
-      <ThemedText
-        style={tailwind("self-center text-sm")}
-        light={tailwind("text-gray-400")}
-        dark={tailwind("text-gray-500")}
-        testID="oracle_price_percentage"
-      >
-        {translate(
-          "screens/CompositeSwapScreen",
-          "Oracle price {{percentageChange}}",
-          {
-            percentageChange: oraclePriceText,
-          }
-        )}
-      </ThemedText>
-      <View style={tailwind("flex flex-row items-center")}>
-        <Icon height={20} width={20} />
-        <ThemedText style={tailwind("pl-2")}>{tokenDisplaySymbol}</ThemedText>
-      </View>
-    </ThemedView>
-  );
-}
-
-// Separated from TokenRow due to custom UI styling difficulties
-interface TargetTokenForm {
-  token: TokenState | OwnedTokenState;
-  control: Control<{ tokenA: string; tokenB: string }>;
-}
-
-function TargetTokenRow(form: TargetTokenForm): JSX.Element {
-  const { token, control } = form;
-  const Icon = getNativeIcon(token.displaySymbol);
-  const rules: {
-    required: boolean;
-    pattern: RegExp;
-    validate: any;
-    max?: string;
-  } = {
-    required: true,
-    pattern: /^\d*\.?\d*$/,
-    validate: {
-      greaterThanZero: (value: string) =>
-        new BigNumber(
-          value !== undefined && value !== "" ? value : 0
-        ).isGreaterThan(0),
-    },
-  };
-  const defaultValue = "";
-
-  return (
-    <Controller
-      control={control}
-      defaultValue={defaultValue}
-      name="tokenB"
-      render={({ field: { value } }) => (
-        <ThemedView
-          light={tailwind("bg-gray-50")}
-          style={tailwind(
-            "flex-row flex-grow justify-between items-center p-2 rounded"
-          )}
-        >
-          <NumberFormat
-            value={value}
-            thousandSeparator
-            displayType="text"
-            renderText={(value) => (
-              <ThemedText
-                style={tailwind("self-center text-sm")}
-                light={tailwind("text-gray-500")}
-                dark={tailwind("text-gray-400")}
-                testID="text_input_tokenB"
-              >
-                {value}
-              </ThemedText>
-            )}
-          />
-          <View style={tailwind("flex flex-row items-center")}>
-            <Icon height={20} width={20} />
-            <ThemedText style={tailwind("pl-2")}>
-              {token.displaySymbol}
-            </ThemedText>
-          </View>
-        </ThemedView>
-      )}
-      rules={rules}
-    />
   );
 }
