+++ mobile-app/app/screens/AppNavigator/screens/Dex/CompositeSwap/components/SlippageTolerance.tsx
@@ -1,25 +1,42 @@
-import { ReactElement, useEffect, useState } from "react";
+import {
+  ThemedIcon,
+  ThemedTextV2,
+  ThemedTouchableOpacityV2,
+  ThemedViewV2,
+} from "@components/themed";
 import BigNumber from "bignumber.js";
 import { tailwind } from "@tailwind";
 import { translate } from "@translations";
-import {
-  ThemedIcon,
-  ThemedText,
-  ThemedTouchableOpacity,
-  ThemedView,
-} from "@components/themed";
-import { View } from "@components";
-import { BottomSheetInfo } from "@components/BottomSheetInfo";
-import { WalletTextInput } from "@components/WalletTextInput";
-import { debounce } from "lodash";
+import { useEffect, useState } from "react";
+import { StyleProp, View, ViewStyle } from "react-native";
+import { WalletTextInputV2 } from "@components/WalletTextInputV2";
 
-const percentageList = ["0.5", "1", "3", "5", "10"];
+export interface SlippageError {
+  type: "error" | "helper";
+  text?: string;
+  style?: StyleProp<ViewStyle>;
+}
 
-interface SlippageToleranceProps {
+interface SlippageToleranceCardProps {
   slippage: BigNumber;
   slippageError?: SlippageError;
   setSlippageError: (error?: SlippageError) => void;
-  setSlippage: (slippage: BigNumber) => void;
+  setSlippage: (val: BigNumber) => void;
+  onPress: () => void;
+  isEditing: boolean;
+  setIsEditing: (val: boolean) => void;
+}
+
+export enum SlippageAmountButtonTypes {
+  ZeroPointFive = "0.5",
+  One = "1",
+  Three = "3",
+}
+
+export enum TransactionCardStatus {
+  Default,
+  Active,
+  Error,
 }
 
 export function SlippageTolerance({
@@ -27,94 +44,30 @@ export function SlippageTolerance({
   setSlippage,
   slippageError,
   setSlippageError,
-}: SlippageToleranceProps): JSX.Element {
-  const [isCustomSlippage, setIsCustomSlippage] = useState(false);
-  const slippageTolerance = {
-    title: "Slippage tolerance",
-    message:
-      "Slippages are rate charges that occur within an order transaction. Note that the slippage tolerance also includes the DEX Stabilization fees. Choose how much of this slippage you are willing to accept.",
-  };
-
-  useEffect(() => {
-    setIsCustomSlippage(
-      !percentageList.includes(new BigNumber(slippage).toString())
-    );
-  }, []);
-
-  return (
-    <View style={tailwind("px-4 pt-4 w-full")}>
-      <View
-        style={tailwind("flex-row items-center")}
-        testID="slippage_tolerance_heading"
-      >
-        <ThemedText
-          dark={tailwind("text-gray-50")}
-          light={tailwind("text-gray-900")}
-          style={tailwind("mr-1")}
-        >
-          {translate("screens/SlippageTolerance", "Slippage tolerance")}
-        </ThemedText>
-        <BottomSheetInfo
-          alertInfo={slippageTolerance}
-          name={slippageTolerance.title}
-        />
-      </View>
-      <SlippageSelector
-        slippage={slippage}
-        slippageError={slippageError}
-        setSlippageError={setSlippageError}
-        isCustomSlippage={isCustomSlippage}
-        setIsCustomSlippage={setIsCustomSlippage}
-        onSubmitSlippage={setSlippage}
-      />
-    </View>
-  );
-}
-
-interface SlippageSelectorProps {
-  isCustomSlippage: boolean;
-  onSubmitSlippage: (val: BigNumber, isCustomSlippage: boolean) => void;
-  setIsCustomSlippage: (val: boolean) => void;
-  slippage: BigNumber;
-  slippageError?: SlippageError;
-  setSlippageError: (error?: SlippageError) => void;
-}
-
-export interface SlippageError {
-  type: "error" | "helper";
-  text?: string;
-}
-
-function SlippageSelector({
-  isCustomSlippage,
-  onSubmitSlippage,
-  slippage,
-  setIsCustomSlippage,
-  slippageError,
-  setSlippageError,
-}: SlippageSelectorProps): JSX.Element {
-  const [selectedSlippage, setSelectedSlippage] = useState(slippage.toString());
+  onPress,
+  isEditing,
+  setIsEditing,
+}: React.PropsWithChildren<SlippageToleranceCardProps>): JSX.Element {
+  const [selectedSlippage, setSelectedSlippage] = useState(slippage.toFixed(8));
   const [isRiskWarningDisplayed, setIsRiskWarningDisplayed] = useState(false);
-  const submitSlippage = debounce(onSubmitSlippage, 500);
 
-  const onSlippageChange = (value: string): void => {
-    setSelectedSlippage(value);
-    submitSlippage(new BigNumber(value), isCustomSlippage);
+  const checkIfCustomSlippage = (slippage: string) => {
+    return !Object.values(SlippageAmountButtonTypes).some((buttonAmount) =>
+      new BigNumber(new BigNumber(buttonAmount).toFixed(8)).isEqualTo(slippage)
+    );
   };
+
+  const [isCustomAmount, setIsCustomAmount] = useState(
+    checkIfCustomSlippage(slippage.toFixed(8))
+  );
 
   const isSlippageValid = (): boolean => {
     return slippageError === undefined || slippageError?.type === "helper";
   };
 
   const validateSlippage = (value: string): void => {
+    /* Allow empty custom slippage -> reverts back to previous slippage */
     if (value === undefined || value === "") {
-      setSlippageError({
-        type: "error",
-        text: translate(
-          "screens/SlippageTolerance",
-          "Required field is missing"
-        ),
-      });
       return;
     } else if (
       !(new BigNumber(value).gte(0) && new BigNumber(value).lte(100))
@@ -128,7 +81,6 @@ function SlippageSelector({
       });
       return;
     }
-
     setSlippageError(undefined);
   };
 
@@ -140,127 +92,258 @@ function SlippageSelector({
     );
   }, [selectedSlippage]);
 
+  const updateSlippage = (slippage: string, updateLocalStorage: boolean) => {
+    // update local storage
+    if (updateLocalStorage) {
+      setSlippage(new BigNumber(slippage));
+    } else {
+      // update local state
+      setSelectedSlippage(slippage);
+    }
+  };
+
+  // updates states when there is a change in slippage from local storage
+  useEffect(() => {
+    // percentage value will be in 2DP
+    updateSlippage(slippage.toFixed(2), false);
+    setIsCustomAmount(checkIfCustomSlippage(slippage.toFixed(2)));
+  }, [slippage]);
+
   return (
     <>
-      <View style={tailwind("flex-row mb-4 flex-wrap")}>
-        {percentageList.map((value) => (
-          <SlippageButton
-            key={value}
-            onPress={() => {
-              setIsCustomSlippage(false);
-              onSlippageChange(value);
-            }}
-            isActive={!isCustomSlippage && selectedSlippage === value}
-            label={`${value}%`}
-          />
-        ))}
-        <SlippageButton
-          onPress={() => {
-            setIsCustomSlippage(true);
-          }}
-          icon={
-            <ThemedIcon
-              size={14}
-              name="edit"
-              iconType="MaterialIcons"
-              style={tailwind("mr-1")}
-              dark={tailwind(
-                `${isCustomSlippage ? "text-darkprimary-500" : "text-gray-50"}`
-              )}
-              light={tailwind(
-                `${isCustomSlippage ? "text-primary-500" : "text-gray-900"}`
-              )}
-            />
-          }
-          isActive={isCustomSlippage}
-          label={translate("screens/SlippageTolerance", "Custom")}
-        />
-      </View>
-
-      {isCustomSlippage && (
-        <ThemedView
-          dark={tailwind("bg-transparent")}
-          light={tailwind("bg-transparent")}
-          style={tailwind("flex-row w-full mb-4")}
+      <View style={tailwind("flex-row items-center pb-2 px-5")}>
+        <ThemedTextV2
+          style={tailwind("text-xs font-normal-v2")}
+          dark={tailwind("text-mono-dark-v2-500")}
+          light={tailwind("text-mono-light-v2-500")}
         >
-          <WalletTextInput
-            onChangeText={onSlippageChange}
-            keyboardType="numeric"
-            autoCapitalize="none"
-            placeholder="0.00%"
-            style={tailwind("flex-grow w-2/5 h-8")}
-            containerStyle="mb-1 w-full flex-col"
-            testID="slippage_input"
-            value={
-              selectedSlippage !== undefined ? selectedSlippage.toString() : ""
-            }
-            displayClearButton={selectedSlippage !== undefined}
-            onClearButtonPress={() => onSlippageChange("")}
-            inputType="numeric"
-            valid={isSlippageValid()}
-            inlineText={slippageError}
+          {translate("screens/CompositeSwapScreen", "SLIPPAGE TOLERANCE")}
+        </ThemedTextV2>
+        <ThemedTouchableOpacityV2
+          style={tailwind("pl-1")}
+          onPress={onPress}
+          testID="slippage_info_button"
+        >
+          <ThemedIcon
+            name="info-outline"
+            iconType="MaterialIcons"
+            size={16}
+            dark={tailwind("text-mono-dark-v2-900")}
+            light={tailwind("text-mono-light-v2-900")}
           />
-        </ThemedView>
+        </ThemedTouchableOpacityV2>
+      </View>
+      {isEditing ? (
+        <>
+          <View style={tailwind("flex-row")}>
+            <View style={tailwind("flex-row items-center mr-2 w-9/12")}>
+              <WalletTextInputV2
+                onChangeText={(val: string) => {
+                  updateSlippage(val, false);
+                }}
+                keyboardType="numeric"
+                autoCapitalize="none"
+                style={tailwind("flex-grow w-2/5 font-normal-v2 text-xs")}
+                inputContainerStyle={tailwind("h-9")}
+                testID="slippage_input"
+                value={selectedSlippage !== undefined ? selectedSlippage : ""}
+                displayClearButton={selectedSlippage !== ""}
+                onClearButtonPress={() => {
+                  updateSlippage("", false);
+                  setSlippageError(undefined);
+                }}
+                inputType="numeric"
+                inlineText={slippageError}
+                valid={isSlippageValid()}
+                helperContainerStyle={tailwind("mx-5")}
+                borderContainerStyle={tailwind("rounded-2xl-v2")}
+              />
+            </View>
+            <View style={tailwind("flex-1 h-9")}>
+              <ThemedTouchableOpacityV2
+                light={tailwind("bg-mono-light-v2-900")}
+                dark={tailwind("bg-mono-dark-v2-900")}
+                style={tailwind(
+                  "p-2.5 justify-center items-center flex-grow rounded-full z-10",
+                  {
+                    "opacity-30": !isSlippageValid(),
+                  }
+                )}
+                onPress={() => {
+                  setIsEditing(false);
+                  if (
+                    selectedSlippage === undefined ||
+                    selectedSlippage === ""
+                  ) {
+                    updateSlippage(slippage.toFixed(2), false); // reset to previous slippage
+                  } else {
+                    setIsCustomAmount(true);
+                    updateSlippage(
+                      new BigNumber(selectedSlippage).toFixed(2),
+                      true
+                    );
+                  }
+                }}
+                disabled={!isSlippageValid()}
+                testID="set_slippage_button"
+              >
+                <ThemedTextV2
+                  light={tailwind("text-mono-light-v2-100")}
+                  dark={tailwind("text-mono-dark-v2-100")}
+                  style={tailwind("text-xs font-semibold-v2")}
+                >
+                  {translate("components/CompositeSwapScreen", "Set")}
+                </ThemedTextV2>
+              </ThemedTouchableOpacityV2>
+            </View>
+          </View>
+        </>
+      ) : (
+        <ThemedViewV2
+          light={tailwind("bg-mono-light-v2-00")}
+          dark={tailwind("bg-mono-dark-v2-00")}
+          style={tailwind(
+            "flex flex-row justify-around items-center rounded-full"
+          )}
+        >
+          {Object.values(SlippageAmountButtonTypes).map((type) => {
+            return (
+              <PercentageAmountButton
+                key={type}
+                onPress={() => {
+                  updateSlippage(type, true);
+                  setIsCustomAmount(false);
+                }}
+                percentageAmount={type}
+                isSelected={
+                  !isCustomAmount &&
+                  new BigNumber(selectedSlippage).isEqualTo(
+                    new BigNumber(type).toFixed(2)
+                  )
+                }
+              />
+            );
+          })}
+          <CustomAmountButton
+            setIsCustomSlippage={() => {
+              setIsEditing(true);
+            }}
+            isCustomAmount={isCustomAmount}
+            customAmount={new BigNumber(selectedSlippage).toFixed(2)}
+          />
+        </ThemedViewV2>
       )}
-
       {isRiskWarningDisplayed && (
-        <ThemedText
-          light={tailwind("text-error-500")}
-          dark={tailwind("text-darkerror-500")}
-          style={tailwind("text-sm mb-3")}
+        <ThemedTextV2
+          light={tailwind("text-warning-600")}
+          dark={tailwind("text-darkwarning-600")}
+          style={tailwind("font-normal-v2 text-xs mt-2 mx-5")}
           testID="slippage_warning"
         >
-          {translate("screens/SlippageTolerance", "Proceed at your own risk")}
-        </ThemedText>
+          {translate(
+            "screens/SlippageTolerance",
+            "Set high tolerance at your own risk"
+          )}
+        </ThemedTextV2>
       )}
     </>
   );
 }
 
-function SlippageButton({
-  onPress,
-  isActive,
-  label,
-  icon,
-}: {
+interface PercentageAmountButtonProps {
   onPress: () => void;
-  isActive: boolean;
-  label: string;
-  icon?: ReactElement;
-}): JSX.Element {
-  const buttonStyles = "flex flex-row px-2 py-1.5 rounded items-center";
-  const activeStyle = "text-primary-500";
+  percentageAmount: SlippageAmountButtonTypes;
+  isSelected: boolean;
+}
+
+function PercentageAmountButton({
+  onPress,
+  percentageAmount,
+  isSelected,
+}: PercentageAmountButtonProps): JSX.Element {
   return (
-    <View style={tailwind("mr-1.5 mt-2")}>
-      <ThemedTouchableOpacity
-        key={label}
-        onPress={onPress}
-        style={tailwind(`${buttonStyles} ${isActive ? activeStyle : ""} `)}
-        testID={`slippage_${label}`}
-        light={tailwind({
-          "bg-primary-50": isActive,
-          "bg-gray-50": !isActive,
+    <ThemedTouchableOpacityV2
+      light={tailwind({ "bg-mono-light-v2-900": isSelected })}
+      dark={tailwind({ "bg-mono-dark-v2-900": isSelected })}
+      style={tailwind(
+        "w-3/12 items-center rounded-full justify-center self-stretch h-9"
+      )}
+      onPress={onPress}
+      testID={`slippage_${percentageAmount}%`}
+    >
+      <ThemedTextV2
+        light={tailwind("text-mono-light-v2-700", {
+          "text-mono-light-v2-100": isSelected,
         })}
-        dark={tailwind({
-          "bg-darkprimary-50": isActive,
-          "bg-gray-900": !isActive,
+        dark={tailwind("text-mono-dark-v2-700", {
+          "text-mono-dark-v2-100": isSelected,
         })}
+        style={tailwind("text-xs px-4 py-2.5 font-normal-v2")}
       >
-        {icon}
-        <ThemedText
-          dark={tailwind(
-            `${
-              isActive ? "text-darkprimary-500 font-semibold" : "text-gray-50"
-            }`
-          )}
-          light={tailwind(
-            `${isActive ? "text-primary-500 font-semibold" : "text-gray-900"}`
-          )}
-          style={tailwind("text-sm")}
+        {percentageAmount}%
+      </ThemedTextV2>
+    </ThemedTouchableOpacityV2>
+  );
+}
+
+interface CustomAmountButtonProps {
+  isCustomAmount: boolean;
+  customAmount: string;
+  setIsCustomSlippage: () => void;
+}
+
+function CustomAmountButton({
+  isCustomAmount,
+  customAmount,
+  setIsCustomSlippage,
+}: CustomAmountButtonProps): JSX.Element {
+  return (
+    <ThemedTouchableOpacityV2
+      light={tailwind("bg-transparent", {
+        "bg-mono-light-v2-900": isCustomAmount,
+      })}
+      dark={tailwind("bg-transparent", {
+        "bg-mono-dark-v2-900": isCustomAmount,
+      })}
+      style={tailwind("rounded-r-full w-3/12 items-center", {
+        "rounded-l-full": isCustomAmount,
+      })}
+      onPress={setIsCustomSlippage}
+      testID="slippage_custom"
+    >
+      <ThemedViewV2
+        light={tailwind("text-mono-light-v2-500")}
+        dark={tailwind("text-mono-dark-v2-500")}
+        style={tailwind(
+          "font-semibold-v2 text-xs px-4 py-2.5 flex-row items-center h-9"
+        )}
+      >
+        <ThemedTextV2
+          light={tailwind("text-mono-light-v2-500", {
+            "text-mono-light-v2-100": isCustomAmount,
+          })}
+          dark={tailwind("text-mono-dark-v2-500", {
+            "text-mono-dark-v2-100": isCustomAmount,
+          })}
+          style={tailwind("text-xs font-normal-v2", {
+            "pr-1.5": isCustomAmount,
+          })}
+          testID="slippage_custom_amount"
         >
-          {label}
-        </ThemedText>
-      </ThemedTouchableOpacity>
-    </View>
+          {isCustomAmount
+            ? `${customAmount}%`
+            : translate("screens/CompositeSwapScreen", "Custom")}
+        </ThemedTextV2>
+        {isCustomAmount && (
+          <ThemedIcon
+            name="edit-2"
+            size={16}
+            iconType="Feather"
+            dark={tailwind("text-mono-dark-v2-100")}
+            light={tailwind("text-mono-light-v2-100")}
+          />
+        )}
+      </ThemedViewV2>
+    </ThemedTouchableOpacityV2>
   );
 }
