+++ mobile-app/app/screens/AppNavigator/screens/Dex/CompositeSwap/components/SwapSummary.tsx
@@ -9,7 +9,7 @@ import { ThemedIcon, ThemedTextV2, ThemedViewV2 } from "@components/themed";
 import { PriceRateProps, PricesSectionV2 } from "@components/PricesSectionV2";
 import { BottomSheetInfoV2 } from "@components/BottomSheetInfoV2";
 import { NumberRowV2 } from "@components/NumberRowV2";
-import { ButtonGroupTabKey } from "../CompositeSwapScreen";
+import { ButtonGroupTabKey } from "@screens/AppNavigator/screens/Dex/CompositeSwap/components/SwapButtonGroup";
 import { DexStabilizationType } from "../../hook/DexStabilization";
 
 interface SwapSummaryProps {
