+++ mobile-app/app/screens/AppNavigator/screens/Portfolio/screens/WithdrawFutureSwapScreen.tsx
@@ -1,33 +1,46 @@
 import { useEffect, useMemo, useState } from "react";
-import { Platform, StyleProp, ViewStyle } from "react-native";
 import BigNumber from "bignumber.js";
-import Slider from "@react-native-community/slider";
 import { StackScreenProps } from "@react-navigation/stack";
 import { tailwind } from "@tailwind";
 import { translate } from "@translations";
-import { View } from "@components";
 import { getNativeIcon } from "@components/icons/assets";
 import {
   ThemedIcon,
-  ThemedScrollView,
-  ThemedText,
-  ThemedTouchableOpacity,
-  ThemedView,
+  ThemedScrollViewV2,
+  ThemedSectionTitleV2,
+  ThemedTextV2,
+  ThemedTouchableOpacityV2,
+  ThemedViewV2,
 } from "@components/themed";
-import { WalletTextInput } from "@components/WalletTextInput";
-import { NumberRow } from "@components/NumberRow";
-import { InfoRow, InfoType } from "@components/InfoRow";
 import { useWhaleApiClient } from "@shared-contexts/WhaleContext";
 import { useLogger } from "@shared-contexts/NativeLoggingProvider";
-import { SubmitButtonGroup } from "@components/SubmitButtonGroup";
 import { hasTxQueued } from "@store/transaction_queue";
 import { hasTxQueued as hasBroadcastQueued } from "@store/ocean";
 import { RootState } from "@store";
 import { useSelector } from "react-redux";
 import { NavigationProp, useNavigation } from "@react-navigation/native";
+import { ViewPoolHeader } from "@screens/AppNavigator/screens/Dex/components/ViewPoolHeader";
+import {
+  AmountButtonTypes,
+  TransactionCard,
+  TransactionCardStatus,
+} from "@components/TransactionCard";
+import { WalletTransactionCardTextInput } from "@components/WalletTransactionCardTextInput";
+import { useToast } from "react-native-toast-notifications";
+import { InputHelperTextV2 } from "@components/InputHelperText";
+import { NumberRowV2 } from "@components/NumberRowV2";
+import { ButtonV2 } from "@components/ButtonV2";
+import { useBottomSheet } from "@hooks/useBottomSheet";
+import { useThemeContext } from "@shared-contexts/ThemeProvider";
+import { ViewSwapInfo } from "@screens/AppNavigator/screens/Portfolio/screens/ViewSwapInfo";
+import { Platform, View } from "react-native";
+import {
+  BottomSheetWebWithNavV2,
+  BottomSheetWithNavV2,
+} from "@components/BottomSheetWithNavV2";
+import { useFutureSwapDate } from "../../Dex/hook/FutureSwap";
 import { useTokenPrice } from "../hooks/TokenPrice";
 import { PortfolioParamList } from "../PortfolioNavigator";
-import { useFutureSwapDate } from "../../Dex/hook/FutureSwap";
 
 type Props = StackScreenProps<PortfolioParamList, "WithdrawFutureSwapScreen">;
 
@@ -39,6 +52,7 @@ export function WithdrawFutureSwapScreen(props: Props): JSX.Element {
   const navigation = useNavigation<NavigationProp<PortfolioParamList>>();
   const client = useWhaleApiClient();
   const logger = useLogger();
+  const { isLight } = useThemeContext();
   const { getTokenPrice } = useTokenPrice();
   const hasPendingJob = useSelector((state: RootState) =>
     hasTxQueued(state.transactionQueue)
@@ -47,9 +61,24 @@ export function WithdrawFutureSwapScreen(props: Props): JSX.Element {
     hasBroadcastQueued(state.ocean)
   );
   const blockCount = useSelector((state: RootState) => state.block.count ?? 0);
-  const { isEnded } = useFutureSwapDate(executionBlock, blockCount);
-  const [amountToWithdraw, setAmountToWithdraw] = useState("0");
+  const { isEnded, transactionDate } = useFutureSwapDate(
+    executionBlock,
+    blockCount
+  );
+  const [amountToWithdraw, setAmountToWithdraw] = useState("");
   const [fee, setFee] = useState<BigNumber>(new BigNumber(0.0001));
+
+  const toast = useToast();
+  const TOAST_DURATION = 2000;
+  const [transactionCardStatus, setTransactionCardStatus] =
+    useState<TransactionCardStatus>();
+  const [isInputFocus, setIsInputFocus] = useState<boolean>(false);
+  const [hasError, setHasError] = useState<boolean>(false);
+  const modalSortingSnapPoints = {
+    ios: ["45%"],
+    android: ["45%"],
+  };
+  const IconSource = getNativeIcon(source.displaySymbol);
 
   useEffect(() => {
     client.fee
@@ -57,6 +86,20 @@ export function WithdrawFutureSwapScreen(props: Props): JSX.Element {
       .then((f) => setFee(new BigNumber(f)))
       .catch(logger.error);
   }, []);
+
+  useEffect(() => {
+    setTransactionCardStatus(
+      hasError
+        ? TransactionCardStatus.Error
+        : isInputFocus
+        ? TransactionCardStatus.Active
+        : TransactionCardStatus.Default
+    );
+  }, [hasError, isInputFocus]);
+
+  useEffect(() => {
+    setHasError(BigNumber(amountToWithdraw).gt(BigNumber(source.amount)));
+  }, [source, amountToWithdraw]);
 
   const { remainingAmount, remainingAmountInUSD } = useMemo(() => {
     return {
@@ -71,9 +114,52 @@ export function WithdrawFutureSwapScreen(props: Props): JSX.Element {
     };
   }, [source, amountToWithdraw]);
 
-  const onAmountChange = (amount: string): void => {
-    setAmountToWithdraw(amount);
+  const {
+    bottomSheetRef,
+    containerRef,
+    expandModal,
+    dismissModal,
+    isModalDisplayed,
+  } = useBottomSheet();
+
+  const bottomSheetHeader = {
+    headerStatusBarHeight: 2,
+    headerTitle: "",
+    headerBackTitleVisible: false,
+    headerStyle: tailwind("rounded-t-xl-v2 border-b-0", {
+      "bg-mono-light-v2-100": isLight,
+      "bg-mono-dark-v2-100": !isLight,
+    }),
+    headerRight: (): JSX.Element => {
+      return (
+        <ThemedTouchableOpacityV2
+          style={tailwind("mr-5")}
+          onPress={dismissModal}
+          testID="close_bottom_sheet_button"
+        >
+          <ThemedIcon iconType="Feather" name="x-circle" size={22} />
+        </ThemedTouchableOpacityV2>
+      );
+    },
+    headerLeft: () => <></>,
   };
+
+  const viewSwapInfo = useMemo(() => {
+    return [
+      {
+        stackScreenName: "ViewSwapInfo",
+        component: ViewSwapInfo({
+          futureSwap: {
+            source: source,
+            destination: destination,
+          },
+          executionBlock: executionBlock,
+          transactionDate: transactionDate,
+        }),
+        option: bottomSheetHeader,
+      },
+    ];
+  }, [isLight]);
 
   const onSubmit = async (): Promise<void> => {
     navigation.navigate({
@@ -99,296 +185,204 @@ export function WithdrawFutureSwapScreen(props: Props): JSX.Element {
     });
   };
 
-  const rowStyle = {
-    lhsThemedProps: {
-      light: tailwind("text-gray-500"),
-      dark: tailwind("text-gray-400"),
-    },
-    rhsThemedProps: {
-      light: tailwind("text-gray-900"),
-      dark: tailwind("text-gray-50"),
-    },
-  };
+  function onPercentagePress(amount: string, type: AmountButtonTypes): void {
+    setAmountToWithdraw(amount);
+    showToast(type);
+  }
+
+  function showToast(type: AmountButtonTypes): void {
+    if (source === undefined) {
+      return;
+    }
+
+    toast.hideAll();
+    const isMax = type === AmountButtonTypes.Max;
+    const toastMessage = isMax
+      ? "Max available {{unit}} entered"
+      : "{{percent}} of available {{unit}} entered";
+    const toastOption = {
+      unit: source.displaySymbol,
+      percent: type,
+    };
+    toast.show(
+      translate("screens/WithdrawFutureSwapScreen", toastMessage, toastOption),
+      {
+        type: "wallet_toast",
+        placement: "top",
+        duration: TOAST_DURATION,
+      }
+    );
+  }
 
   return (
-    <ThemedScrollView>
-      <View style={tailwind("flex flex-row mt-3 mx-2")}>
-        <TokenSelection
-          label={translate("screens/WithdrawFutureSwapScreen", "FROM")}
-          symbol={source.displaySymbol}
-        />
-        <TokenSelection
-          label={translate("screens/WithdrawFutureSwapScreen", "TO")}
-          symbol={destination.displaySymbol}
-        />
-      </View>
-
-      <AmountSlider
-        current={Number(amountToWithdraw)}
-        maxValue={Number(source.amount)}
-        onChange={onAmountChange}
-        viewStyle={tailwind("p-4 mx-4 mt-4 rounded-t flex flex-col")}
-      />
-      <ThemedView
-        dark={tailwind("bg-gray-800")}
-        light={tailwind("bg-white")}
-        style={tailwind("flex-row px-4 pb-4 mx-4 items-stretch rounded-b")}
+    <View ref={containerRef} style={tailwind("flex-col flex-1")}>
+      <ThemedScrollViewV2
+        contentContainerStyle={tailwind("flex-grow justify-between pb-12 pt-8")}
       >
-        <WalletTextInput
-          inputType="numeric"
-          onChange={(event) => {
-            setAmountToWithdraw(event.nativeEvent.text);
-          }}
-          testID="text_input_percentage"
-          placeholder={translate(
-            "screens/WithdrawFutureSwapScreen",
-            "Enter an amount"
-          )}
-          value={amountToWithdraw}
-          keyboardType="numeric"
-          containerStyle="w-full"
-          style={tailwind("flex-grow")}
-          inlineText={{
-            type: "helper",
-            text: (
-              <HelperText
-                displayedValue={
-                  new BigNumber(amountToWithdraw).isNaN()
-                    ? "0"
-                    : amountToWithdraw
-                }
-              />
-            ),
-          }}
-          displayClearButton
-        >
-          <TokenIcon icon={source.displaySymbol} />
-        </WalletTextInput>
-      </ThemedView>
-
-      {!new BigNumber(amountToWithdraw).isNaN() &&
-        !new BigNumber(amountToWithdraw).isZero() && (
-          <ThemedView
-            dark={tailwind("bg-gray-800")}
-            light={tailwind("bg-white")}
-            style={tailwind("m-4 items-stretch rounded-b")}
-          >
-            <NumberRow
-              lhs={translate(
+        <View>
+          <ViewPoolHeader
+            tokenASymbol={source.displaySymbol}
+            tokenBSymbol={destination.displaySymbol}
+            headerLabel={translate(
+              "screens/WithdrawFutureSwapScreen",
+              "View swap info"
+            )}
+            onPress={expandModal}
+            testID="view_swap_info"
+          />
+          <View style={tailwind("flex-col mt-2 px-5")}>
+            <ThemedSectionTitleV2
+              testID="withdraw_title"
+              text={translate(
                 "screens/WithdrawFutureSwapScreen",
-                "Amount to withdraw"
-              )}
-              rhs={{
-                value: new BigNumber(amountToWithdraw).toFixed(8),
-                testID: "text_amount_to_withdraw",
-                suffixType: "text",
-                suffix: source.displaySymbol,
-              }}
-              lhsThemedProps={rowStyle.lhsThemedProps}
-              rhsThemedProps={rowStyle.rhsThemedProps}
-              rhsUsdAmount={getTokenPrice(
-                source.symbol,
-                new BigNumber(amountToWithdraw),
-                false
+                "I WANT TO WITHDRAW"
               )}
             />
-            <NumberRow
-              lhs={translate(
-                "screens/WithdrawFutureSwapScreen",
-                "Remaining amount"
-              )}
-              rhs={{
-                value: remainingAmount,
-                testID: "text_remaining_amount",
-                suffixType: "text",
-                suffix: source.displaySymbol,
-              }}
-              rhsUsdAmount={remainingAmountInUSD}
-              lhsThemedProps={rowStyle.lhsThemedProps}
-              rhsThemedProps={rowStyle.rhsThemedProps}
-            />
-            <InfoRow
-              type={InfoType.EstimatedFee}
-              value={fee.toFixed(8)}
-              testID="text_fee"
-              suffix="DFI"
+            <TransactionCard
+              maxValue={new BigNumber(source.amount)}
+              onChange={onPercentagePress}
+              amountButtonsStyle={{ style: tailwind("border-t-0.5") }}
               containerStyle={{
-                style: tailwind("rounded-b p-4 flex-row items-start w-full"),
-                dark: tailwind("bg-gray-800"),
-                light: tailwind("bg-white"),
+                style: tailwind("px-5 rounded-t-lg-v2"),
               }}
-              lhsThemedProps={rowStyle.lhsThemedProps}
-              rhsThemedProps={rowStyle.rhsThemedProps}
-            />
-          </ThemedView>
-        )}
-
-      <View style={tailwind("mb-2")}>
-        <SubmitButtonGroup
-          isDisabled={
-            hasPendingJob ||
-            hasPendingBroadcastJob ||
-            new BigNumber(amountToWithdraw).isNaN() ||
-            new BigNumber(amountToWithdraw).isZero() ||
-            isEnded
-          }
-          label={translate("screens/WithdrawFutureSwapScreen", "CONTINUE")}
-          processingLabel={translate(
-            "screens/WithdrawFutureSwapScreen",
-            "CONTINUE"
-          )}
-          onSubmit={onSubmit}
-          title="submit"
-          isProcessing={hasPendingJob || hasPendingBroadcastJob}
-          displayCancelBtn={false}
-        />
-      </View>
-    </ThemedScrollView>
-  );
-}
-
-function TokenSelection(props: {
-  symbol?: string;
-  label: string;
-}): JSX.Element {
-  const Icon = getNativeIcon(props.symbol ?? "");
-  return (
-    <View style={[tailwind("flex-grow mx-2"), { flexBasis: 0 }]}>
-      <ThemedText
-        dark={tailwind("text-gray-400")}
-        light={tailwind("text-gray-500")}
-        style={tailwind("text-xs pb-1")}
-      >
-        {props.label}
-      </ThemedText>
-      <ThemedTouchableOpacity
-        testID={`token_select_button_${props.label}`}
-        dark={tailwind("bg-gray-600 text-gray-500 border-0")}
-        light={tailwind("bg-gray-200 border-0")}
-        style={tailwind("flex flex-row items-center border rounded p-3")}
-        disabled
-      >
-        {props.symbol === undefined && (
-          <ThemedText
-            dark={tailwind("text-gray-400")}
-            light={tailwind("text-gray-500")}
-            style={tailwind("text-sm leading-6")}
-          >
-            {translate("screens/WithdrawFutureSwapScreen", "Select token")}
-          </ThemedText>
-        )}
-
-        {props.symbol !== undefined && (
-          <>
-            <Icon testID="tokenA_icon" height={17} width={17} />
-            <ThemedText
-              style={tailwind("ml-2")}
-              dark={tailwind("text-gray-400")}
-              light={tailwind("text-gray-500")}
+              status={transactionCardStatus}
             >
-              {props.symbol}
-            </ThemedText>
-          </>
+              <ThemedViewV2
+                light={tailwind("border-mono-light-v2-300")}
+                dark={tailwind("border-mono-dark-v2-300")}
+                style={tailwind("flex flex-row items-center py-4.5")}
+              >
+                <IconSource height={20} width={20} />
+                <WalletTransactionCardTextInput
+                  onFocus={() => setIsInputFocus(true)}
+                  onBlur={() => setIsInputFocus(false)}
+                  onChangeText={setAmountToWithdraw}
+                  placeholder="0.00"
+                  value={amountToWithdraw}
+                  inputType="numeric"
+                  displayClearButton={amountToWithdraw !== ""}
+                  onClearButtonPress={() => setAmountToWithdraw("")}
+                  testID="text_input_percentage"
+                  containerStyle="flex-1 flex-col"
+                  inputContainerStyle={tailwind("pl-2 pr-0 py-0")}
+                />
+              </ThemedViewV2>
+            </TransactionCard>
+            {hasError ? (
+              <ThemedTextV2
+                style={tailwind("font-normal-v2 text-xs px-5 pt-2 text-red-v2")}
+                light={tailwind("text-red-v2")}
+                dark={tailwind("text-red-v2")}
+                testID="text_inline_error"
+              >
+                {translate(
+                  "screens/WithdrawFutureSwapScreen",
+                  "Insufficient Balance"
+                )}
+              </ThemedTextV2>
+            ) : (
+              <View style={tailwind("pt-1 px-1")}>
+                <InputHelperTextV2
+                  testID="text_inline"
+                  label={`${translate(
+                    "screens/WithdrawFutureSwapScreen",
+                    "Withdraw from"
+                  )}: `}
+                  content={source.amount}
+                  suffix={` ${source.displaySymbol}`}
+                />
+              </View>
+            )}
+          </View>
+
+          {amountToWithdraw.length > 0 && (
+            <View style={tailwind("flex-col")}>
+              <ThemedViewV2
+                style={tailwind("rounded-lg-v2 border-0.5 mt-8 mx-5 p-5 pb-0")}
+                light={tailwind("border-mono-light-v2-300")}
+                dark={tailwind("border-mono-dark-v2-300")}
+              >
+                <NumberRowV2
+                  lhs={{
+                    value: translate(
+                      "screens/WithdrawFutureSwapScreen",
+                      "Transaction fee"
+                    ),
+                    testID: "transaction_fee_label",
+                    themedProps: {
+                      light: tailwind("text-mono-light-v2-500"),
+                      dark: tailwind("text-mono-dark-v2-500"),
+                    },
+                  }}
+                  rhs={{
+                    value: fee.toFixed(8),
+                    testID: "text_fee",
+                    suffix: " DFI",
+                    usdAmount: getTokenPrice("DFI", new BigNumber(fee)),
+                    usdTextStyle: tailwind("text-sm"),
+                    usdContainerStyle: tailwind("pt-1"),
+                  }}
+                />
+              </ThemedViewV2>
+              <ThemedTextV2
+                style={tailwind(
+                  "font-normal-v2 text-xs text-center pt-12 pb-5"
+                )}
+                light={tailwind("text-mono-light-v2-500")}
+                dark={tailwind("text-mono-dark-v2-500")}
+              >
+                {translate(
+                  "screens/WithdrawFutureSwapScreen",
+                  "Review full details in the next screen"
+                )}
+              </ThemedTextV2>
+            </View>
+          )}
+        </View>
+
+        <View style={tailwind("w-full px-12")}>
+          <ButtonV2
+            fillType="fill"
+            label={translate("components/Button", "Continue")}
+            disabled={
+              hasPendingJob ||
+              hasPendingBroadcastJob ||
+              new BigNumber(amountToWithdraw).isNaN() ||
+              new BigNumber(amountToWithdraw).isZero() ||
+              isEnded ||
+              hasError
+            }
+            styleProps="w-full"
+            onPress={onSubmit}
+            testID="button_continue_withdraw"
+          />
+        </View>
+
+        {Platform.OS === "web" ? (
+          <BottomSheetWebWithNavV2
+            modalRef={containerRef}
+            screenList={viewSwapInfo}
+            isModalDisplayed={isModalDisplayed}
+            // eslint-disable-next-line react-native/no-inline-styles
+            modalStyle={{
+              position: "absolute",
+              bottom: "0",
+              height: "404px",
+              width: "375px",
+              zIndex: 50,
+              borderTopLeftRadius: 15,
+              borderTopRightRadius: 15,
+              overflow: "hidden",
+            }}
+          />
+        ) : (
+          <BottomSheetWithNavV2
+            modalRef={bottomSheetRef}
+            screenList={viewSwapInfo}
+            snapPoints={modalSortingSnapPoints}
+          />
         )}
-        <ThemedIcon
-          iconType="MaterialIcons"
-          name="unfold-more"
-          size={20}
-          dark={tailwind("text-gray-400")}
-          light={tailwind("text-gray-500")}
-          style={[tailwind("text-center mt-0.5"), { marginLeft: "auto" }]}
-        />
-      </ThemedTouchableOpacity>
+      </ThemedScrollViewV2>
     </View>
-  );
-}
-
-function AmountSlider(props: {
-  current: number;
-  maxValue: number;
-  onChange: (percentage: string) => void;
-  viewStyle: StyleProp<ViewStyle>;
-}): JSX.Element {
-  return (
-    <ThemedView
-      dark={tailwind("bg-gray-800")}
-      light={tailwind("bg-white")}
-      style={props.viewStyle}
-    >
-      <Slider
-        maximumValue={props.maxValue}
-        minimumTrackTintColor="#ff00af"
-        minimumValue={0}
-        onSlidingComplete={(val) =>
-          props.onChange(new BigNumber(val).toFixed(8))
-        }
-        testID="slider_future_swap_screen"
-        thumbTintColor="#ff00af"
-        value={isNaN(props.current) ? 0 : props.current}
-        style={tailwind({
-          "-mb-2": Platform.OS !== "web",
-        })}
-      />
-      <View style={tailwind("flex flex-row justify-between mt-1")}>
-        <ThemedText
-          style={tailwind("text-xs")}
-          dark={tailwind("text-gray-400")}
-          light={tailwind("text-gray-500")}
-        >
-          0
-        </ThemedText>
-        <ThemedText
-          style={tailwind("text-xs")}
-          dark={tailwind("text-gray-400")}
-          light={tailwind("text-gray-500")}
-        >
-          {props.maxValue.toFixed(8)}
-        </ThemedText>
-      </View>
-    </ThemedView>
-  );
-}
-
-function HelperText(props: { displayedValue: string }): JSX.Element {
-  return (
-    <ThemedView
-      light={tailwind("bg-transparent")}
-      dark={tailwind("bg-transparent")}
-      style={tailwind("flex-row mt-2")}
-    >
-      <ThemedText
-        style={tailwind("text-sm")}
-        dark={tailwind("text-gray-400")}
-        light={tailwind("text-gray-500")}
-      >
-        {`${translate(
-          "screens/WithdrawFutureSwapScreen",
-          "You are withdrawing"
-        )} `}
-      </ThemedText>
-      <ThemedText
-        style={tailwind("text-sm font-semibold")}
-        dark={tailwind("text-white")}
-        light={tailwind("text-black")}
-        testID="displayed_withdraw_amount"
-      >
-        {new BigNumber(props.displayedValue).toFixed(8)}
-      </ThemedText>
-    </ThemedView>
-  );
-}
-
-function TokenIcon(props: { icon: string }): JSX.Element {
-  const Icon = getNativeIcon(props.icon);
-
-  return (
-    <ThemedView
-      dark={tailwind("bg-gray-800")}
-      light={tailwind("bg-white")}
-      style={tailwind("flex-row items-center mr-2")}
-    >
-      <Icon width={17} height={17} />
-      <ThemedText style={tailwind("ml-1")}>{props.icon}</ThemedText>
-    </ThemedView>
   );
 }
